refactor: address review feedback on domain tool packaging
Move recorder, weather, and response-service tools into domain modules and built-in packages so the MCP server no longer owns their tool surface.
