+++ custom_components/mcp_assist/config_flow.py
@@ -150,9 +150,6 @@ from .const import (
     TOOL_FAMILY_EXTERNAL_CUSTOM,
     TOOL_FAMILY_MEMORY,
     TOOL_FAMILY_PROFILE_SETTINGS,
-    TOOL_FAMILY_RECORDER,
-    TOOL_FAMILY_RESPONSE_SERVICE,
-    TOOL_FAMILY_WEATHER_FORECAST,
     TOOL_FAMILY_SHARED_SETTINGS,
     OPENAI_BASE_URL,
     OPENROUTER_BASE_URL,
@@ -302,9 +299,6 @@ STATIC_TOOL_FAMILY_ALPHABETICAL = [
     TOOL_FAMILY_EXTERNAL_CUSTOM,
     TOOL_FAMILY_DEVICE,
     TOOL_FAMILY_MEMORY,
-    TOOL_FAMILY_RECORDER,
-    TOOL_FAMILY_RESPONSE_SERVICE,
-    TOOL_FAMILY_WEATHER_FORECAST,
 ]
 
 PROFILE_DISABLE_FIELD_BY_FAMILY = {
@@ -312,9 +306,6 @@ PROFILE_DISABLE_FIELD_BY_FAMILY = {
     TOOL_FAMILY_EXTERNAL_CUSTOM: DISABLE_CUSTOM_TOOLS_FIELD,
     TOOL_FAMILY_DEVICE: DISABLE_DEVICE_FIELD,
     TOOL_FAMILY_MEMORY: DISABLE_MEMORY_FIELD,
-    TOOL_FAMILY_RECORDER: DISABLE_RECORDER_FIELD,
-    TOOL_FAMILY_RESPONSE_SERVICE: DISABLE_RESPONSE_SERVICE_FIELD,
-    TOOL_FAMILY_WEATHER_FORECAST: DISABLE_WEATHER_FORECAST_FIELD,
 }
 
 STATIC_TOOL_FAMILY_SHARED_LABELS = {
@@ -322,9 +313,6 @@ STATIC_TOOL_FAMILY_SHARED_LABELS = {
     TOOL_FAMILY_EXTERNAL_CUSTOM: "Custom Tools",
     TOOL_FAMILY_DEVICE: "Device Tools",
     TOOL_FAMILY_MEMORY: "Memory",
-    TOOL_FAMILY_RECORDER: "Recorder History",
-    TOOL_FAMILY_RESPONSE_SERVICE: "Response Service Reads",
-    TOOL_FAMILY_WEATHER_FORECAST: "Weather Forecast",
 }
 
 STATIC_TOOL_FAMILY_PROFILE_DISABLE_LABELS = {
@@ -332,9 +320,6 @@ STATIC_TOOL_FAMILY_PROFILE_DISABLE_LABELS = {
     TOOL_FAMILY_EXTERNAL_CUSTOM: "Disable Custom Tools",
     TOOL_FAMILY_DEVICE: "Disable Device Tools",
     TOOL_FAMILY_MEMORY: "Disable Memory",
-    TOOL_FAMILY_RECORDER: "Disable Recorder History",
-    TOOL_FAMILY_RESPONSE_SERVICE: "Disable Response Service Reads",
-    TOOL_FAMILY_WEATHER_FORECAST: "Disable Weather Forecast",
 }
