+++ custom_components/mcp_assist/mcp_server.py
@@ -2,7 +2,7 @@
 
 import asyncio
 import base64
-from collections import Counter, defaultdict
+from collections import defaultdict
 import ipaddress
 import json
 import logging
@@ -11,7 +11,7 @@ from pathlib import Path, PurePosixPath
 import re
 from typing import Any, Dict, List, Tuple
 from urllib.parse import urlparse
-from datetime import date, datetime, time, timedelta
+from datetime import timedelta
 
 import aiohttp
 from aiohttp import web, WSMsgType
@@ -21,7 +21,7 @@ from voluptuous_openapi import convert
 import yarl
 
 from homeassistant.components import conversation
-from homeassistant.core import Context, HomeAssistant, SupportsResponse
+from homeassistant.core import Context, HomeAssistant
 from homeassistant.exceptions import HomeAssistantError
 from homeassistant.helpers import (
     area_registry as ar,
@@ -29,10 +29,8 @@ from homeassistant.helpers import (
     entity_registry as er,
     llm,
     network as network_helper,
-    service as service_helper,
 )
 from homeassistant.components.homeassistant import async_should_expose
-from homeassistant.components.recorder import history
 from homeassistant.util import dt as dt_util
 
 try:
@@ -47,9 +45,9 @@ except ImportError:  # pragma: no cover - older Home Assistant versions
 
 from .custom_tools.builtin_catalog import (
     BuiltInToolToggleSpec,
+    get_builtin_toggle_spec_by_package_id,
     is_builtin_package_enabled_for_shared_settings,
 )
-from .custom_tools.music_assistant import summarize_music_assistant_response
 from .const import (
     DOMAIN,
     MCP_SERVER_NAME,
@@ -96,7 +94,6 @@ from .domain_registry import (
     validate_domain_action,
     validate_service_parameters,
     get_supported_domains,
-    get_domain_info,
     get_domains_by_type,
     TYPE_CONTROLLABLE,
     TYPE_READ_ONLY,
@@ -106,6 +103,10 @@ from .provider_runtime import (
     build_provider_auth_headers,
     resolve_provider_runtime_config,
 )
+from .server_tools.calendar import CalendarToolsMixin
+from .server_tools.recorder import RecorderToolsMixin
+from .server_tools.response_services import ResponseServicesMixin
+from .server_tools.weather import WeatherToolsMixin
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -114,7 +115,12 @@ _HTTP_REDIRECT_STATUSES = {301, 302, 303, 307, 308}
 _MAX_IMAGE_FETCH_REDIRECTS = 5
 
 
-class MCPServer:
+class MCPServer(
+    CalendarToolsMixin,
+    RecorderToolsMixin,
+    ResponseServicesMixin,
+    WeatherToolsMixin,
+):
     """MCP Server for entity discovery."""
 
     def __init__(self, hass: HomeAssistant, port: int, entry=None) -> None:
@@ -243,12 +249,13 @@ class MCPServer:
 
     def _weather_forecast_tool_enabled(self) -> bool:
         """Return whether weather forecast MCP helpers are enabled."""
+        built_in_spec = get_builtin_toggle_spec_by_package_id(
+            "weather_forecast",
+            self._get_builtin_toggle_specs(),
+        )
+        if built_in_spec is not None:
+            return self._is_builtin_package_enabled(built_in_spec)
         return bool(
-            self._get_shared_setting(
-                CONF_ENABLE_RESPONSE_SERVICE_TOOLS,
-                DEFAULT_ENABLE_RESPONSE_SERVICE_TOOLS,
-            )
-        ) and bool(
             self._get_shared_setting(
                 CONF_ENABLE_WEATHER_FORECAST_TOOL,
                 DEFAULT_ENABLE_WEATHER_FORECAST_TOOL,
@@ -266,6 +273,12 @@ class MCPServer:
 
     def _response_service_tools_enabled(self) -> bool:
         """Return whether native response-service tools are enabled."""
+        built_in_spec = get_builtin_toggle_spec_by_package_id(
+            "response_service",
+            self._get_builtin_toggle_specs(),
+        )
+        if built_in_spec is not None:
+            return self._is_builtin_package_enabled(built_in_spec)
         return bool(
             self._get_shared_setting(
                 CONF_ENABLE_RESPONSE_SERVICE_TOOLS,
@@ -275,6 +288,12 @@ class MCPServer:
 
     def _recorder_tools_enabled(self) -> bool:
         """Return whether recorder/history tools are enabled."""
+        built_in_spec = get_builtin_toggle_spec_by_package_id(
+            "recorder",
+            self._get_builtin_toggle_specs(),
+        )
+        if built_in_spec is not None:
+            return self._is_builtin_package_enabled(built_in_spec)
         return bool(
             self._get_shared_setting(
                 CONF_ENABLE_RECORDER_TOOLS,
@@ -1669,217 +1688,6 @@ class MCPServer:
                 },
             },
             {
-                "name": "get_weather_forecast",
-                "description": "Get a Home Assistant weather forecast in one call. Prefer this for user weather questions before web search. It finds the weather entity, chooses a supported forecast type, and summarizes today or tomorrow.",
-                "llmDescription": "Get a Home Assistant weather forecast.",
-                "inputSchema": {
-                    "$schema": "http://json-schema.org/draft-07/schema#",
-                    "type": "object",
-                    "properties": {
-                        "entity_id": {
-                            "type": "string",
-                            "description": "Optional specific weather entity ID to use.",
-                        },
-                        "area": {
-                            "type": "string",
-                            "description": "Optional area or room name if you have multiple weather entities.",
-                        },
-                        "floor": {
-                            "type": "string",
-                            "description": "Optional floor name if you have multiple weather entities.",
-                        },
-                        "label": {
-                            "type": "string",
-                            "description": "Optional label name if you have multiple weather entities.",
-                        },
-                        "name_contains": {
-                            "type": "string",
-                            "description": "Optional text to match a specific weather entity name or alias.",
-                        },
-                        "when": {
-                            "type": "string",
-                            "description": "Forecast day to summarize. Use 'today', 'tomorrow', or a local date like '2026-04-13'. Defaults to 'tomorrow'.",
-                        },
-                        "forecast_type": {
-                            "type": "string",
-                            "enum": ["daily", "twice_daily", "hourly"],
-                            "description": "Optional forecast type override. If unsupported by the target entity, the tool falls back to a supported type.",
-                        },
-                    },
-                    "required": [],
-                    "additionalProperties": False,
-                },
-            },
-            {
-                "name": "get_calendar_events",
-                "description": "Get upcoming Home Assistant calendar events in one call. Prefer this for questions like 'When is the next Mariners game?' or 'What's on our calendar tomorrow?'. It discovers matching calendars, calls calendar.get_events, and summarizes the next matching event or agenda.",
-                "llmDescription": "Get upcoming Home Assistant calendar events, schedules, or subscribed team games.",
-                "routingHints": {
-                    "preferred_when": "Use for calendar, schedule, or next-game questions when Home Assistant may already have the answer.",
-                    "example_queries": [
-                        "When is the next Mariners game?",
-                    ],
-                },
-                "inputSchema": {
-                    "$schema": "http://json-schema.org/draft-07/schema#",
-                    "type": "object",
-                    "properties": {
-                        "entity_id": {
-                            "type": "string",
-                            "description": "Optional specific calendar entity ID to use.",
-                        },
-                        "area": {
-                            "type": "string",
-                            "description": "Optional area or room name if you want calendars associated with a specific area.",
-                        },
-                        "floor": {
-                            "type": "string",
-                            "description": "Optional floor name if you want calendars associated with a specific floor.",
-                        },
-                        "label": {
-                            "type": "string",
-                            "description": "Optional label name if you want calendars tagged with a specific label.",
-                        },
-                        "query": {
-                            "type": "string",
-                            "description": "Optional general text to match either calendar names or event details, for example 'Mariners' or 'dentist'.",
-                        },
-                        "name_contains": {
-                            "type": "string",
-                            "description": "Optional text to match a specific calendar entity name or alias.",
-                        },
-                        "event_text": {
-                            "type": "string",
-                            "description": "Optional text to match event summary, title, description, or location.",
-                        },
-                        "when": {
-                            "type": "string",
-                            "description": "Optional time window anchor. Use 'now', 'today', 'tomorrow', a local date like '2026-04-13', or an ISO datetime.",
-                        },
-                        "days": {
-                            "type": "integer",
-                            "description": "How many days forward to search when 'when' is omitted or when searching from a specific start time. Defaults to 60 for upcoming searches and 1 for day-specific lookups.",
-                            "default": 60,
-                            "minimum": 1,
-                            "maximum": 365,
-                        },
-                        "limit": {
-                            "type": "integer",
-                            "description": "Maximum number of matching events to summarize.",
-                            "default": 5,
-                            "minimum": 1,
-                            "maximum": 20,
-                        },
-                    },
-                    "required": [],
-                    "additionalProperties": False,
-                },
-            },
-            {
-                "name": "list_response_services",
-                "description": "List Home Assistant services that currently support native response data. Use this when you need to discover which read/query-style services can be called with call_service_with_response.",
-                "llmDescription": "List services that return structured response data.",
-                "inputSchema": {
-                    "$schema": "http://json-schema.org/draft-07/schema#",
-                    "type": "object",
-                    "properties": {
-                        "domain": {
-                            "type": "string",
-                            "description": "Optional domain filter, for example 'weather', 'calendar', or 'media_player'.",
-                        },
-                        "query": {
-                            "type": "string",
-                            "description": "Optional text filter matching domain, service, name, or description.",
-                        },
-                        "limit": {
-                            "type": "integer",
-                            "description": "Maximum number of services to return (default: 50, max: 200).",
-                            "default": 50,
-                            "minimum": 1,
-                            "maximum": 200,
-                        },
-                    },
-                    "required": [],
-                    "additionalProperties": False,
-                },
-            },
-            {
-                "name": "call_service_with_response",
-                "description": "Call a Home Assistant service that returns structured response data for read/query use cases. Use this for native service-response reads like calendar or to-do queries, media browsing/searching, or integration-specific query data. For normal weather questions, prefer get_weather_forecast.",
-                "llmDescription": "Call a Home Assistant service that returns structured data.",
-                "inputSchema": {
-                    "$schema": "http://json-schema.org/draft-07/schema#",
-                    "type": "object",
-                    "properties": {
-                        "domain": {
-                            "type": "string",
-                            "description": "The Home Assistant domain for the response-returning service, for example 'weather', 'calendar', 'todo', or 'media_player'.",
-                        },
-                        "service": {
-                            "type": "string",
-                            "description": "The service/action name to call, for example 'get_forecasts', 'get_events', 'get_items', 'browse_media', or 'search_media'.",
-                        },
-                        "target": {
-                            "type": "object",
-                            "description": "Optional target entities or selector IDs such as areas, floors, labels, or devices. Selectors are resolved to exposed entity IDs, and may be narrowed using the service's target metadata when available.",
-                            "properties": {
-                                "entity_id": {
-                                    "oneOf": [
-                                        {"type": "string"},
-                                        {"type": "array", "items": {"type": "string"}},
-                                    ],
-                                    "description": "Single entity ID or list of entity IDs.",
-                                },
-                                "area_id": {
-                                    "oneOf": [
-                                        {"type": "string"},
-                                        {"type": "array", "items": {"type": "string"}},
-                                    ],
-                                    "description": "Single area ID or list of area IDs.",
-                                },
-                                "floor_id": {
-                                    "oneOf": [
-                                        {"type": "string"},
-                                        {"type": "array", "items": {"type": "string"}},
-                                    ],
-                                    "description": "Single floor ID or list of floor IDs.",
-                                },
-                                "label_id": {
-                                    "oneOf": [
-                                        {"type": "string"},
-                                        {"type": "array", "items": {"type": "string"}},
-                                    ],
-                                    "description": "Single label ID or list of label IDs.",
-                                },
-                                "device_id": {
-                                    "oneOf": [
-                                        {"type": "string"},
-                                        {"type": "array", "items": {"type": "string"}},
-                                    ],
-                                    "description": "Single device ID or list of device IDs.",
-                                },
-                            },
-                            "minProperties": 1,
-                            "additionalProperties": False,
-                        },
-                        "data": {
-                            "type": "object",
-                            "description": "Additional service data for the response-returning service. Required fields are validated from Home Assistant's native service descriptions when available.",
-                            "additionalProperties": True,
-                        },
-                        "timeout": {
-                            "type": "integer",
-                            "description": "Timeout in seconds (default: 60, max: 300).",
-                            "default": 60,
-                            "minimum": 1,
-                            "maximum": 300,
-                        },
-                    },
-                    "required": ["domain", "service"],
-                    "additionalProperties": False,
-                },
-            },
-            {
                 "name": "set_conversation_state",
                 "description": "Indicate whether you expect a response from the user after your message",
                 "inputSchema": {
@@ -1944,120 +1752,6 @@ class MCPServer:
                         },
                     },
                     "required": ["automation_id"],
-                    "additionalProperties": False,
-                },
-            },
-            {
-                "name": "get_entity_history",
-                "description": "Get recorder-backed history for a specific entity. By default this returns a recent timeline, and with mode='last_event' it returns only the most recent matching event or change.",
-                "llmDescription": "Get recorder-backed history for an entity.",
-                "inputSchema": {
-                    "$schema": "http://json-schema.org/draft-07/schema#",
-                    "type": "object",
-                    "properties": {
-                        "entity_id": {
-                            "type": "string",
-                            "description": "The entity ID to get history for. Use discover_entities first and pass the discovered entity ID here.",
-                        },
-                        "mode": {
-                            "type": "string",
-                            "enum": ["timeline", "last_event"],
-                            "default": "timeline",
-                            "description": "timeline returns recent history entries; last_event returns only the most recent matching event or change.",
-                        },
-                        "event": {
-                            "type": "string",
-                            "description": "Optional semantic event filter for last_event mode, such as opened, closed, on, off, locked, unlocked, detected, cleared, home, or away.",
-                        },
-                        "state": {
-                            "oneOf": [
-                                {"type": "string"},
-                                {
-                                    "type": "array",
-                                    "items": {"type": "string"},
-                                },
-                            ],
-                            "description": "Optional target state or states to filter by. Works with both timeline and last_event modes.",
-                        },
-                        "hours": {
-                            "type": "integer",
-                            "description": "Number of hours of recorder history to search. In timeline mode the default is 24 hours; in last_event mode the default is 720 hours (30 days). Max: 8760 hours / 1 year.",
-                            "minimum": 1,
-                            "maximum": 8760,
-                        },
-                        "limit": {
-                            "type": "integer",
-                            "description": "Maximum number of timeline entries to return in timeline mode (default: 50, max: 100). Most recent changes shown first.",
-                            "default": 50,
-                            "minimum": 1,
-                            "maximum": 100,
-                        },
-                    },
-                    "required": ["entity_id"],
-                    "additionalProperties": False,
-                },
-            },
-            {
-                "name": "analyze_entity_history",
-                "description": "Analyze Home Assistant recorder history for aggregate questions such as 'how many times was the door opened in the last hour?', 'how long has it been locked?', or 'how often did this sensor trigger today?'. Can count all changes or matching states/events.",
-                "llmDescription": "Analyze recorder history for counts, durations, or matching transitions.",
-                "inputSchema": {
-                    "$schema": "http://json-schema.org/draft-07/schema#",
-                    "type": "object",
-                    "properties": {
-                        "entity_id": {
-                            "type": "string",
-                            "description": "The entity ID to analyze in recorder history.",
-                        },
-                        "event": {
-                            "type": "string",
-                            "description": "Optional semantic event to analyze, such as opened, closed, on, off, locked, unlocked, detected, cleared, home, or away.",
-                        },
-                        "state": {
-                            "oneOf": [
-                                {"type": "string"},
-                                {
-                                    "type": "array",
-                                    "items": {"type": "string"},
-                                },
-                            ],
-                            "description": "Optional specific target state or states to analyze. If omitted, counts all recorded state changes in the time window.",
-                        },
-                        "hours": {
-                            "type": "integer",
-                            "description": "How far back to analyze recorder history. Default is 24 hours for count/summary/duration/stats, and 720 hours (30 days) for streak. Max: 8760 hours / 1 year.",
-                            "minimum": 1,
-                            "maximum": 8760,
-                        },
-                        "analysis": {
-                            "type": "string",
-                            "enum": ["count", "summary", "duration", "streak", "stats"],
-                            "default": "count",
-                            "description": "count returns how many matching events occurred; summary also includes the first and last matching times in the window; duration reports total time spent in the matching state; streak reports how long the entity has continuously been in the matching state right now; stats reports numeric min/max/average over the window.",
-                        },
-                    },
-                    "required": ["entity_id"],
-                    "additionalProperties": False,
-                },
-            },
-            {
-                "name": "get_entity_state_at_time",
-                "description": "Look up the recorder state of an entity at a specific date/time. Use this for questions like 'was the gate open at 2 PM?' or 'what was the temperature at 9 this morning?'",
-                "llmDescription": "Look up an entity's recorder state at a specific time.",
-                "inputSchema": {
-                    "$schema": "http://json-schema.org/draft-07/schema#",
-                    "type": "object",
-                    "properties": {
-                        "entity_id": {
-                            "type": "string",
-                            "description": "The entity ID to inspect in recorder history.",
-                        },
-                        "datetime": {
-                            "type": "string",
-                            "description": "The target date/time to inspect, preferably as an ISO 8601 timestamp. If no timezone is included, Home Assistant local time is assumed.",
-                        },
-                    },
-                    "required": ["entity_id", "datetime"],
                     "additionalProperties": False,
                 },
             },
@@ -2199,28 +1893,12 @@ class MCPServer:
             return await self.tool_get_assist_context_snapshot(arguments)
         elif tool_name == "perform_action":
             return await self.tool_perform_action(arguments)
-        elif tool_name == "get_weather_forecast":
-            return await self.tool_get_weather_forecast(arguments)
-        elif tool_name == "get_calendar_events":
-            return await self.tool_get_calendar_events(arguments)
-        elif tool_name == "list_response_services":
-            return await self.tool_list_response_services(arguments)
-        elif tool_name == "call_service_with_response":
-            return await self.tool_call_service_with_response(arguments)
         elif tool_name == "set_conversation_state":
             return await self.tool_set_conversation_state(arguments)
         elif tool_name == "run_script":
             return await self.tool_run_script(arguments)
         elif tool_name == "run_automation":
             return await self.tool_run_automation(arguments)
-        elif tool_name == "get_entity_history":
-            return await self.tool_get_entity_history(arguments)
-        elif tool_name == "get_last_entity_event":
-            return await self.tool_get_last_entity_event(arguments)
-        elif tool_name == "analyze_entity_history":
-            return await self.tool_analyze_entity_history(arguments)
-        elif tool_name == "get_entity_state_at_time":
-            return await self.tool_get_entity_state_at_time(arguments)
         elif tool_name == "remember_memory":
             return await self.tool_remember_memory(arguments)
         elif tool_name == "recall_memories":
@@ -3792,287 +3470,6 @@ class MCPServer:
             ]
         }
 
-    async def tool_get_weather_forecast(self, args: Dict[str, Any]) -> Dict[str, Any]:
-        """Get a Home Assistant weather forecast in one call."""
-        target_entity_id = str(args.get("entity_id") or "").strip() or None
-        area = str(args.get("area") or "").strip() or None
-        floor = str(args.get("floor") or "").strip() or None
-        label = str(args.get("label") or "").strip() or None
-        name_contains = str(args.get("name_contains") or "").strip() or None
-        when_value = str(args.get("when") or "tomorrow").strip() or "tomorrow"
-        forecast_type = str(args.get("forecast_type") or "").strip().lower() or None
-        timeout = self._coerce_int_arg(
-            args.get("timeout"), default=60, minimum=1, maximum=300
-        )
-
-        target_date, day_label, parse_error = self._parse_weather_forecast_when(
-            when_value
-        )
-        if parse_error:
-            return {
-                "content": [{"type": "text", "text": f"❌ Error: {parse_error}"}]
-            }
-
-        try:
-            resolved_target, entity_info = await self._resolve_weather_forecast_target(
-                entity_id=target_entity_id,
-                area=area,
-                floor=floor,
-                label=label,
-                name_contains=name_contains,
-            )
-        except ValueError as err:
-            return {"content": [{"type": "text", "text": f"❌ Error: {err}"}]}
-
-        prepared_data = self._prepare_response_service_data(
-            "weather",
-            "get_forecasts",
-            {"type": forecast_type} if forecast_type else {},
-            resolved_target=resolved_target,
-        )
-
-        service_result = await self.tool_call_service_with_response(
-            {
-                "domain": "weather",
-                "service": "get_forecasts",
-                "target": resolved_target,
-                "data": prepared_data,
-                "timeout": timeout,
-            }
-        )
-
-        if service_result.get("response") is None:
-            return service_result
-
-        response = service_result.get("response")
-        if not isinstance(response, dict):
-            return service_result
-
-        payload = response.get(entity_info["entity_id"])
-        forecast_entries = payload.get("forecast") if isinstance(payload, dict) else None
-        if not isinstance(forecast_entries, list):
-            return service_result
-
-        summary_text = self._summarize_requested_weather_forecast(
-            entity_name=entity_info["name"],
-            entity_id=entity_info["entity_id"],
-            forecast_entries=forecast_entries,
-            forecast_type=prepared_data.get("type"),
-            target_date=target_date,
-            day_label=day_label,
-        )
-        if not summary_text:
-            return service_result
-
-        result: Dict[str, Any] = {
-            "content": [{"type": "text", "text": summary_text}],
-            "response": response,
-        }
-        return result
-
-    async def tool_get_calendar_events(self, args: Dict[str, Any]) -> Dict[str, Any]:
-        """Get upcoming calendar events in one call."""
-        target_entity_id = str(args.get("entity_id") or "").strip() or None
-        area = str(args.get("area") or "").strip() or None
-        floor = str(args.get("floor") or "").strip() or None
-        label = str(args.get("label") or "").strip() or None
-        query = str(args.get("query") or "").strip() or None
-        name_contains = str(args.get("name_contains") or "").strip() or None
-        event_text = str(args.get("event_text") or "").strip() or None
-        when_value = str(args.get("when") or "").strip() or None
-        days = self._coerce_int_arg(
-            args.get("days"),
-            default=60,
-            minimum=1,
-            maximum=365,
-        )
-        limit = self._coerce_int_arg(
-            args.get("limit"),
-            default=5,
-            minimum=1,
-            maximum=20,
-        )
-        timeout = self._coerce_int_arg(
-            args.get("timeout"), default=60, minimum=1, maximum=300
-        )
-
-        try:
-            (
-                resolved_target,
-                selected_calendars,
-                fallback_used,
-            ) = await self._resolve_calendar_event_targets(
-                entity_id=target_entity_id,
-                area=area,
-                floor=floor,
-                label=label,
-                query=query,
-                name_contains=name_contains,
-            )
-        except ValueError as err:
-            return self._build_text_tool_result(str(err), is_error=True)
-
-        window_start, window_end, window_label = self._build_calendar_search_window(
-            when_value=when_value,
-            days=days,
-        )
-        if window_start is None or window_end is None:
-            return self._build_text_tool_result(
-                f"Invalid calendar time window: {window_label}",
-                is_error=True,
-            )
-
-        service_result = await self.tool_call_service_with_response(
-            {
-                "domain": "calendar",
-                "service": "get_events",
-                "target": resolved_target,
-                "data": {
-                    "start_date_time": window_start.isoformat(),
-                    "end_date_time": window_end.isoformat(),
-                },
-                "timeout": timeout,
-            }
-        )
-
-        response = service_result.get("response")
-        if not isinstance(response, dict):
-            return service_result
-
-        active_event_filter = event_text
-        if not active_event_filter and fallback_used and query:
-            active_event_filter = query
-
-        matches = self._collect_calendar_event_matches(
-            response=response,
-            selected_calendars=selected_calendars,
-            event_text=active_event_filter,
-        )
-        serialized_matches = self._serialize_service_response_value(matches[:limit])
-        if not matches:
-            filter_label = active_event_filter or name_contains or query or "the request"
-            calendars_label = (
-                f"{len(selected_calendars)} calendar"
-                f"{'' if len(selected_calendars) == 1 else 's'}"
-            )
-            return self._build_text_tool_result(
-                (
-                    f"No upcoming calendar events matched {filter_label!r} in {window_label}. "
-                    f"Searched {calendars_label}."
-                ),
-                structured_content={
-                    "query": query,
-                    "name_contains": name_contains,
-                    "event_text": active_event_filter,
-                    "window": {
-                        "start": window_start.isoformat(),
-                        "end": window_end.isoformat(),
-                        "label": window_label,
-                    },
-                    "selected_calendars": selected_calendars,
-                    "events": [],
-                },
-            )
-
-        summary_text = self._summarize_calendar_matches(
-            matches=matches[:limit],
-            query=query,
-            event_text=active_event_filter,
-            window_label=window_label,
-        )
-        return self._build_text_tool_result(
-            summary_text,
-            structured_content={
-                "query": query,
-                "name_contains": name_contains,
-                "event_text": active_event_filter,
-                "window": {
-                    "start": window_start.isoformat(),
-                    "end": window_end.isoformat(),
-                    "label": window_label,
-                },
-                "selected_calendars": selected_calendars,
-                "events": serialized_matches,
-                "total_matching_events": len(matches),
-            },
-        )
-
-    async def tool_list_response_services(self, args: Dict[str, Any]) -> Dict[str, Any]:
-        """List dynamically available HA services that support response data."""
-        domain_filter = str(args.get("domain") or "").strip().casefold()
-        query = str(args.get("query") or "").strip().casefold()
-        limit = self._coerce_int_arg(
-            args.get("limit"), default=50, minimum=1, maximum=200
-        )
-
-        catalog = await self._get_response_service_catalog()
-        rows: List[tuple[str, str, Dict[str, Any]]] = []
-
-        for domain, services in catalog.items():
-            if domain_filter and domain.casefold() != domain_filter:
-                continue
-
-            for service_name, description in services.items():
-                haystacks = [
-                    domain,
-                    service_name,
-                    str(description.get("name") or ""),
-                    str(description.get("description") or ""),
-                ]
-                if query and not any(query in text.casefold() for text in haystacks):
-                    continue
-
-                rows.append((domain, service_name, description))
-
-        rows.sort(key=lambda item: (item[0], item[1]))
-        rows = rows[:limit]
-
-        if not rows:
-            filters = []
-            if domain_filter:
-                filters.append(f"domain='{domain_filter}'")
-            if query:
-                filters.append(f"query='{query}'")
-            filter_text = f" for {', '.join(filters)}" if filters else ""
-            return {
-                "content": [
-                    {
-                        "type": "text",
-                        "text": f"No response-capable services found{filter_text}.",
-                    }
-                ]
-            }
-
-        text_parts = [
-            f"Found {len(rows)} response-capable Home Assistant services:"
-        ]
-        for domain, service_name, description in rows:
-            detail_parts = [
-                f"response: {description.get('supports_response', 'optional')}"
-            ]
-            target_domains = self._extract_service_target_domains(description)
-            if target_domains:
-                detail_parts.append(f"target domains: {', '.join(target_domains)}")
-            required_fields = self._get_required_service_fields(description)
-            if required_fields:
-                detail_parts.append(f"required: {', '.join(required_fields)}")
-            field_names = self._get_service_field_names(description)
-            if field_names:
-                preview_fields = field_names[:6]
-                detail_parts.append(
-                    "fields: "
-                    + ", ".join(preview_fields)
-                    + ("..." if len(field_names) > len(preview_fields) else "")
-                )
-            if description.get("description"):
-                detail_parts.append(str(description["description"]))
-
-            text_parts.append(
-                f"- {domain}.{service_name} ({'; '.join(detail_parts)})"
-            )
-
-        return {"content": [{"type": "text", "text": "\n".join(text_parts)}]}
-
     async def tool_perform_action(self, args: Dict[str, Any]) -> Dict[str, Any]:
         """Perform an action on Home Assistant entities with progress notifications."""
         domain = args.get("domain")
@@ -4377,160 +3774,6 @@ class MCPServer:
 
             await asyncio.sleep(0.5)
 
-    async def tool_call_service_with_response(
-        self, args: Dict[str, Any]
-    ) -> Dict[str, Any]:
-        """Call a native HA service that returns structured response data."""
-        domain = str(args.get("domain") or "").strip().lower()
-        service = str(args.get("service") or "").strip().lower()
-        target = args.get("target")
-        data = args.get("data", {}) or {}
-        timeout = self._coerce_int_arg(
-            args.get("timeout"), default=60, minimum=1, maximum=300
-        )
-
-        if not domain:
-            return {
-                "content": [
-                    {
-                        "type": "text",
-                        "text": "❌ Error: Missing required parameter 'domain'.",
-                    }
-                ]
-            }
-
-        if not service:
-            return {
-                "content": [
-                    {
-                        "type": "text",
-                        "text": "❌ Error: Missing required parameter 'service'.",
-                    }
-                ]
-            }
-
-        if target is None:
-            target = {}
-        elif not isinstance(target, dict):
-            return {
-                "content": [
-                    {
-                        "type": "text",
-                        "text": "❌ Error: 'target' must be an object with entity_id, area_id, floor_id, label_id, or device_id.",
-                    }
-                ]
-            }
-
-        if not isinstance(data, dict):
-            return {
-                "content": [
-                    {
-                        "type": "text",
-                        "text": "❌ Error: 'data' must be an object of service parameters.",
-                    }
-                ]
-            }
-
-        _LOGGER.info(
-            "📖 Calling response service: %s.%s on %s with data %s",
-            domain,
-            service,
-            target,
-            data,
-        )
-
-        service_description, validation_error = await self._get_response_service_info(
-            domain, service
-        )
-        if validation_error:
-            return {
-                "content": [
-                    {"type": "text", "text": f"❌ Error: {validation_error}"}
-                ]
-            }
-
-        resolved_target = {}
-        try:
-            if target:
-                resolved_target = await self.resolve_target(target)
-                resolved_target = self._restrict_resolved_target_for_service(
-                    resolved_target,
-                    service_description=service_description,
-                )
-        except Exception as err:
-            error_msg = f"Failed to resolve target: {err}"
-            _LOGGER.error(error_msg)
-            return {"content": [{"type": "text", "text": f"❌ Error: {error_msg}"}]}
-
-        prepared_data = self._prepare_response_service_data(
-            domain,
-            service,
-            data,
-            resolved_target=resolved_target,
-        )
-        valid_params, validation_msg = self._validate_response_service_parameters(
-            service_description, prepared_data
-        )
-        if not valid_params:
-            return {
-                "content": [{"type": "text", "text": f"❌ Error: {validation_msg}"}]
-            }
-
-        self.publish_progress(
-            "tool_start",
-            f"Calling response service: {domain}.{service}",
-            tool="call_service_with_response",
-            domain=domain,
-            service=service,
-        )
-
-        try:
-            service_data = {**resolved_target, **prepared_data}
-            response = await asyncio.wait_for(
-                self.hass.services.async_call(
-                    domain=domain,
-                    service=service,
-                    service_data=service_data,
-                    blocking=True,
-                    return_response=True,
-                ),
-                timeout=timeout,
-            )
-        except asyncio.TimeoutError:
-            error_msg = (
-                f"Service-response call timed out after {timeout} seconds: "
-                f"{domain}.{service}"
-            )
-            _LOGGER.error("❌ %s", error_msg)
-            return {"content": [{"type": "text", "text": f"❌ Error: {error_msg}"}]}
-        except Exception as err:
-            error_msg = f"Service-response call failed: {err}"
-            _LOGGER.exception("❌ %s", error_msg)
-            return {"content": [{"type": "text", "text": f"❌ Error: {error_msg}"}]}
-
-        self.publish_progress(
-            "tool_complete",
-            f"Response service completed: {domain}.{service}",
-            tool="call_service_with_response",
-            success=True,
-        )
-
-        serialized_response = self._serialize_service_response_value(response)
-        result_text = self._format_service_response_result(
-            domain,
-            service,
-            resolved_target,
-            serialized_response,
-            request_data=prepared_data,
-        )
-        result: Dict[str, Any] = {
-            "content": [{"type": "text", "text": result_text}]
-        }
-        if serialized_response is not None:
-            result["response"] = serialized_response
-
-        return result
-
     async def tool_set_conversation_state(self, args: Dict[str, Any]) -> Dict[str, Any]:
         """Set whether the assistant expects a user response."""
         expecting_response = args.get("expecting_response", False)
@@ -4668,846 +3911,6 @@ class MCPServer:
             error_msg = f"Automation trigger failed: {err}"
             _LOGGER.exception(f"❌ {error_msg}")
             return {"content": [{"type": "text", "text": f"❌ Error: {error_msg}"}]}
-
-    async def tool_get_entity_history(self, args: Dict[str, Any]) -> Dict[str, Any]:
-        """Get entity history with human-readable formatting."""
-        mode = str(args.get("mode", "timeline")).strip().casefold()
-        if mode not in {"timeline", "last_event"}:
-            return {
-                "content": [
-                    {
-                        "type": "text",
-                        "text": "Invalid mode. Use 'timeline' or 'last_event'.",
-                    }
-                ]
-            }
-        if mode == "last_event":
-            return await self._tool_get_last_entity_event_impl(
-                args,
-                tool_name="get_entity_history",
-            )
-
-        entity_id = args.get("entity_id")
-        hours = self._coerce_int_arg(args.get("hours"), default=24, minimum=1, maximum=8760)
-        limit = self._coerce_int_arg(args.get("limit"), default=50, minimum=1, maximum=100)
-        target_states = self._normalize_history_targets(
-            args.get("state"), args.get("event")
-        )
-
-        _LOGGER.info(f"📜 Getting history for {entity_id}: {hours} hours, limit {limit}")
-
-        # Notify start
-        self.publish_progress(
-            "tool_start",
-            f"Retrieving history for {entity_id}",
-            tool="get_entity_history",
-            entity_id=entity_id,
-        )
-
-        # 1. Get current state
-        current_state = self.hass.states.get(entity_id)
-        if not current_state:
-            return {
-                "content": [
-                    {"type": "text", "text": f"Entity '{entity_id}' not found."}
-                ]
-            }
-
-        friendly_name = current_state.attributes.get("friendly_name", entity_id)
-
-        # 2. Calculate time range (UTC)
-        try:
-            entity_states = await self._fetch_entity_history_states(
-                entity_id,
-                hours=hours,
-                descending=True,
-                limit=None if target_states else limit,
-            )
-        except Exception as e:
-            _LOGGER.error(f"Failed to get history for {entity_id}: {e}")
-            return {
-                "content": [
-                    {"type": "text", "text": f"Failed to retrieve history: {str(e)}"}
-                ]
-            }
-
-        if target_states:
-            entity_states = [
-                state
-                for state in entity_states
-                if state.state.casefold() in target_states
-            ]
-            entity_states = entity_states[:limit]
-
-        # Notify completion
-        self.publish_progress(
-            "tool_complete",
-            f"History retrieved: {len(entity_states)} changes",
-            tool="get_entity_history",
-            success=True,
-        )
-
-        # 4. Format history (most recent first, limited)
-        if not entity_states:
-            if target_states:
-                search_label = self._describe_history_target(
-                    args.get("state"), args.get("event")
-                )
-                return {
-                    "content": [
-                        {
-                            "type": "text",
-                            "text": (
-                                f"{friendly_name} ({entity_id})\n"
-                                f"Current state: {current_state.state}\n\n"
-                                f"No recorded {search_label} entries were found in the last {hours} hours."
-                            ),
-                        }
-                    ]
-                }
-            return {
-                "content": [
-                    {
-                        "type": "text",
-                        "text": f"{friendly_name} ({entity_id})\nCurrent state: {current_state.state}\n\nNo history available for the last {hours} hours.",
-                    }
-                ]
-            }
-
-        # Build formatted text
-        text_parts = [
-            f"{friendly_name} ({entity_id})",
-            f"Current state: {current_state.state}",
-            "",
-            (
-                f"Matching history for {self._describe_history_target(args.get('state'), args.get('event'))} "
-                f"(last {hours} hours):"
-                if target_states
-                else f"Recent history (last {hours} hours):"
-            ),
-        ]
-
-        for state in entity_states:
-            when = state.last_changed or state.last_updated
-            text_parts.append(
-                f"• {self._format_relative_absolute_time(when)} → {state.state}"
-            )
-
-        text_parts.append("")
-        text_parts.append(
-            (
-                f"Showing {len(entity_states)} matching entr{'ies' if len(entity_states) != 1 else 'y'}"
-                if target_states
-                else f"Showing {len(entity_states)} change{'s' if len(entity_states) != 1 else ''}"
-            )
-        )
-
-        return {"content": [{"type": "text", "text": "\n".join(text_parts)}]}
-
-    async def tool_get_last_entity_event(self, args: Dict[str, Any]) -> Dict[str, Any]:
-        """Find the latest recorder event for an entity."""
-        return await self._tool_get_last_entity_event_impl(
-            args,
-            tool_name="get_last_entity_event",
-        )
-
-    async def _tool_get_last_entity_event_impl(
-        self, args: Dict[str, Any], *, tool_name: str
-    ) -> Dict[str, Any]:
-        """Shared implementation for latest-event recorder lookups."""
-        entity_id = args.get("entity_id")
-        hours = self._coerce_int_arg(args.get("hours"), default=720, minimum=1, maximum=8760)
-        current_state = self.hass.states.get(entity_id)
-
-        if not current_state:
-            return {
-                "content": [
-                    {"type": "text", "text": f"Entity '{entity_id}' not found."}
-                ]
-            }
-
-        target_states = self._normalize_history_targets(
-            args.get("state"), args.get("event")
-        )
-        history_candidates = self._history_resolution_candidates(
-            entity_id,
-            args.get("state"),
-            args.get("event"),
-        )
-        selected_candidate = history_candidates[0]
-        history_entity_id = selected_candidate["entity_id"]
-        resolution_note = selected_candidate["note"]
-        current_state = self.hass.states.get(history_entity_id)
-        friendly_name = current_state.attributes.get("friendly_name", history_entity_id)
-        end_time = dt_util.utcnow()
-
-        self.publish_progress(
-            "tool_start",
-            f"Searching recorder history for {history_entity_id}",
-            tool=tool_name,
-            entity_id=history_entity_id,
-        )
-
-        try:
-            if target_states:
-                matched_state = None
-                for candidate in history_candidates:
-                    candidate_entity_id = candidate["entity_id"]
-                    candidate_target_states = target_states
-                    for window_hours in self._build_history_search_windows(hours):
-                        entity_states = await self._fetch_entity_history_states(
-                            candidate_entity_id,
-                            hours=window_hours,
-                            end_time=end_time,
-                            descending=True,
-                        )
-                        candidate_target_states = self._choose_history_count_states(
-                            entity_states,
-                            args.get("state"),
-                            args.get("event"),
-                        )
-                        matched_state = next(
-                            (
-                                state
-                                for state in entity_states
-                                if state.state.casefold() in candidate_target_states
-                            ),
-                            None,
-                        )
-                        if matched_state is not None:
-                            history_entity_id = candidate_entity_id
-                            resolution_note = candidate["note"]
-                            current_state = self.hass.states.get(history_entity_id)
-                            friendly_name = current_state.attributes.get(
-                                "friendly_name", history_entity_id
-                            )
-                            target_states = candidate_target_states
-                            break
-                    if matched_state is not None:
-                        break
-            else:
-                entity_states = await self._fetch_entity_history_states(
-                    history_entity_id,
-                    hours=hours,
-                    end_time=end_time,
-                    descending=True,
-                    limit=1,
-                )
-                matched_state = entity_states[0] if entity_states else None
-        except Exception as err:
-            _LOGGER.error(
-                "Failed to get last recorder event for %s: %s", history_entity_id, err
-            )
-            return {
-                "content": [
-                    {
-                        "type": "text",
-                        "text": f"Failed to retrieve recorder history: {err}",
-                    }
-                ]
-            }
-
-        self.publish_progress(
-            "tool_complete",
-            "Recorder history search complete",
-            tool=tool_name,
-            success=True,
-            found=matched_state is not None,
-        )
-
-        if not matched_state:
-            search_label = self._describe_history_target(
-                args.get("state"), args.get("event")
-            )
-            return {
-                "content": [
-                    {
-                        "type": "text",
-                        "text": (
-                            f"{friendly_name} ({history_entity_id})\n"
-                            f"Current state: {current_state.state}\n\n"
-                            f"No recorded {search_label} event was found in the last {hours} hours."
-                            + (
-                                f"\n\n{resolution_note}"
-                                if resolution_note
-                                else ""
-                            )
-                        ),
-                    }
-                ]
-            }
-
-        when = matched_state.last_changed or matched_state.last_updated
-        search_label = self._describe_history_target(
-            args.get("state"), args.get("event")
-        )
-
-        return {
-            "content": [
-                {
-                    "type": "text",
-                    "text": (
-                        (
-                            f"{resolution_note}\n\n"
-                            if resolution_note
-                            else ""
-                        )
-                        + f"{friendly_name} ({history_entity_id})\n"
-                        f"Current state: {current_state.state}\n"
-                        f"Last recorded {search_label}: {self._format_relative_absolute_time(when)}\n"
-                        f"Matched state: {matched_state.state}"
-                    ),
-                }
-            ]
-        }
-
-    async def tool_analyze_entity_history(self, args: Dict[str, Any]) -> Dict[str, Any]:
-        """Analyze recorder history for counts and summaries."""
-        analysis = str(args.get("analysis", "count")).strip().casefold()
-        if analysis not in {"count", "summary", "duration", "stats", "streak"}:
-            return {
-                "content": [
-                    {
-                        "type": "text",
-                        "text": "Invalid analysis. Use 'count', 'summary', 'duration', 'streak', or 'stats'.",
-                    }
-                ]
-            }
-        entity_id = args.get("entity_id")
-        default_hours = 720 if analysis == "streak" else 24
-        hours = self._coerce_int_arg(
-            args.get("hours"), default=default_hours, minimum=1, maximum=8760
-        )
-        history_candidates = self._history_resolution_candidates(
-            entity_id,
-            args.get("state"),
-            args.get("event"),
-        )
-        selected_candidate = history_candidates[0]
-        history_entity_id = selected_candidate["entity_id"]
-        resolution_note = selected_candidate["note"]
-        current_state = self.hass.states.get(history_entity_id)
-
-        if not current_state:
-            return {
-                "content": [
-                    {"type": "text", "text": f"Entity '{history_entity_id}' not found."}
-                ]
-            }
-
-        friendly_name = current_state.attributes.get("friendly_name", history_entity_id)
-        target_states = self._normalize_history_targets(
-            args.get("state"), args.get("event")
-        )
-
-        self.publish_progress(
-            "tool_start",
-            f"Analyzing recorder history for {history_entity_id}",
-            tool="analyze_entity_history",
-            entity_id=history_entity_id,
-        )
-
-        query_end_time = dt_util.utcnow()
-        query_start_time = query_end_time - timedelta(hours=hours)
-        include_start_time_state = analysis in {"duration", "streak", "stats"}
-
-        async def _load_candidate_history(candidate: Dict[str, Any]) -> Dict[str, Any] | None:
-            candidate_entity_id = candidate["entity_id"]
-            candidate_current_state = self.hass.states.get(candidate_entity_id)
-            if candidate_current_state is None:
-                return None
-
-            candidate_entity_states = await self._fetch_entity_history_states(
-                candidate_entity_id,
-                hours=hours,
-                end_time=query_end_time,
-                descending=False,
-                include_start_time_state=include_start_time_state,
-            )
-            if target_states:
-                candidate_target_filter_states = self._choose_history_count_states(
-                    candidate_entity_states,
-                    args.get("state"),
-                    args.get("event"),
-                )
-            else:
-                candidate_target_filter_states = []
-
-            return {
-                "candidate": candidate,
-                "entity_id": candidate_entity_id,
-                "current_state": candidate_current_state,
-                "friendly_name": candidate_current_state.attributes.get(
-                    "friendly_name", candidate_entity_id
-                ),
-                "entity_states": candidate_entity_states,
-                "target_filter_states": candidate_target_filter_states,
-            }
-
-        try:
-            candidate_history = await _load_candidate_history(selected_candidate)
-        except Exception as err:
-            _LOGGER.error(
-                "Failed to analyze history for %s: %s", history_entity_id, err
-            )
-            return {
-                "content": [
-                    {
-                        "type": "text",
-                        "text": f"Failed to analyze recorder history: {err}",
-                    }
-                ]
-            }
-
-        if candidate_history is None:
-            return {
-                "content": [
-                    {"type": "text", "text": f"Entity '{history_entity_id}' not found."}
-                ]
-            }
-
-        history_entity_id = candidate_history["entity_id"]
-        current_state = candidate_history["current_state"]
-        friendly_name = candidate_history["friendly_name"]
-        entity_states = candidate_history["entity_states"]
-
-        target_filter_states = candidate_history["target_filter_states"]
-
-        if analysis == "duration":
-            if not target_filter_states:
-                return {
-                    "content": [
-                        {
-                            "type": "text",
-                            "text": (
-                                f"{friendly_name} ({history_entity_id})\n"
-                                "Duration analysis requires a target event or state, such as opened, closed, on, or off."
-                            ),
-                        }
-                    ]
-                }
-
-            duration_result = self._calculate_state_duration(
-                entity_states,
-                target_filter_states,
-                query_start_time,
-                query_end_time,
-            )
-            if duration_result["interval_count"] == 0 and len(history_candidates) > 1:
-                for candidate in history_candidates[1:]:
-                    candidate_history = await _load_candidate_history(candidate)
-                    if candidate_history is None:
-                        continue
-                    candidate_duration_result = self._calculate_state_duration(
-                        candidate_history["entity_states"],
-                        candidate_history["target_filter_states"],
-                        query_start_time,
-                        query_end_time,
-                    )
-                    candidate_current_state = candidate_history["current_state"]
-                    if (
-                        candidate_duration_result["interval_count"] > 0
-                        or (
-                            candidate_history["target_filter_states"]
-                            and candidate_current_state.state.casefold()
-                            in candidate_history["target_filter_states"]
-                        )
-                    ):
-                        history_entity_id = candidate_history["entity_id"]
-                        current_state = candidate_current_state
-                        friendly_name = candidate_history["friendly_name"]
-                        entity_states = candidate_history["entity_states"]
-                        target_filter_states = candidate_history["target_filter_states"]
-                        resolution_note = candidate["note"]
-                        duration_result = candidate_duration_result
-                        break
-
-            total_duration = duration_result["total_duration"]
-            interval_count = duration_result["interval_count"]
-            search_label = self._describe_history_target(
-                args.get("state"), args.get("event")
-            )
-
-            self.publish_progress(
-                "tool_complete",
-                "Recorder history duration analysis complete",
-                tool="analyze_entity_history",
-                success=True,
-                seconds=int(total_duration.total_seconds()),
-            )
-
-            text_parts = [
-                f"{friendly_name} ({history_entity_id})",
-                f"Current state: {current_state.state}",
-                f"Total time in {search_label} state during the last {hours} hour{'s' if hours != 1 else ''}: {self._format_duration(total_duration)}",
-                f"Matching interval{'s' if interval_count != 1 else ''}: {interval_count}",
-            ]
-            text_parts = self._prepend_resolution_note(text_parts, resolution_note)
-
-            if interval_count:
-                first_start = duration_result.get("first_start")
-                last_end = duration_result.get("last_end")
-                if first_start is not None:
-                    text_parts.append(
-                        "First matching interval in window started: "
-                        f"{self._format_relative_absolute_time(first_start)}"
-                    )
-                if last_end is not None and last_end < query_end_time:
-                    text_parts.append(
-                        f"Last matching interval ended: {self._format_relative_absolute_time(last_end)}"
-                    )
-
-            if current_state.state.casefold() in target_filter_states:
-                streak_start = current_state.last_changed or current_state.last_updated
-                text_parts.append(
-                    f"Current ongoing {search_label} streak: {self._format_duration(query_end_time - streak_start)}"
-                )
-
-            if not interval_count:
-                text_parts.append("No matching recorder intervals were found in that window.")
-
-            return {"content": [{"type": "text", "text": "\n".join(text_parts)}]}
-
-        if analysis == "streak":
-            if not target_filter_states:
-                return {
-                    "content": [
-                        {
-                            "type": "text",
-                            "text": (
-                                f"{friendly_name} ({history_entity_id})\n"
-                                "Streak analysis requires a target event or state, such as locked, opened, on, or home."
-                            ),
-                        }
-                    ]
-                }
-
-            search_label = self._describe_history_target(
-                args.get("state"), args.get("event")
-            )
-            if (
-                current_state.state.casefold() not in target_filter_states
-                and len(history_candidates) > 1
-            ):
-                for candidate in history_candidates[1:]:
-                    candidate_history = await _load_candidate_history(candidate)
-                    if candidate_history is None:
-                        continue
-                    candidate_current_state = candidate_history["current_state"]
-                    if (
-                        candidate_history["target_filter_states"]
-                        and candidate_current_state.state.casefold()
-                        in candidate_history["target_filter_states"]
-                    ):
-                        history_entity_id = candidate_history["entity_id"]
-                        current_state = candidate_current_state
-                        friendly_name = candidate_history["friendly_name"]
-                        entity_states = candidate_history["entity_states"]
-                        target_filter_states = candidate_history["target_filter_states"]
-                        resolution_note = candidate["note"]
-                        break
-
-            if current_state.state.casefold() not in target_filter_states:
-                text_parts = [
-                    f"{friendly_name} ({history_entity_id})",
-                    f"Current state: {current_state.state}",
-                    f"It is not currently {search_label}, so there is no ongoing {search_label} streak to measure.",
-                ]
-                text_parts = self._prepend_resolution_note(text_parts, resolution_note)
-                return {"content": [{"type": "text", "text": "\n".join(text_parts)}]}
-
-            streak_start = query_start_time
-            exact_start = False
-            streak_reaches_window_start = True
-
-            for window_hours in self._build_history_search_windows(hours):
-                window_start = query_end_time - timedelta(hours=window_hours)
-                window_states = await self._fetch_entity_history_states(
-                    history_entity_id,
-                    hours=window_hours,
-                    end_time=query_end_time,
-                    descending=False,
-                    include_start_time_state=True,
-                )
-                if not window_states:
-                    continue
-                if window_states[-1].state.casefold() not in target_filter_states:
-                    break
-
-                start_idx = len(window_states) - 1
-                while (
-                    start_idx > 0
-                    and window_states[start_idx - 1].state.casefold() in target_filter_states
-                ):
-                    start_idx -= 1
-
-                streak_start = (
-                    window_states[start_idx].last_changed
-                    or window_states[start_idx].last_updated
-                )
-                if start_idx > 0:
-                    exact_start = True
-                    streak_reaches_window_start = False
-                    break
-                if streak_start > window_start:
-                    exact_start = True
-                    streak_reaches_window_start = False
-                    break
-                streak_reaches_window_start = True
-                if window_hours >= hours:
-                    break
-
-            streak_duration = query_end_time - max(streak_start, query_start_time)
-
-            self.publish_progress(
-                "tool_complete",
-                "Recorder history streak analysis complete",
-                tool="analyze_entity_history",
-                success=True,
-                seconds=int(streak_duration.total_seconds()),
-            )
-
-            duration_text = self._format_duration(streak_duration)
-            text_parts = [
-                f"{friendly_name} ({history_entity_id})",
-                f"Current state: {current_state.state}",
-            ]
-            if exact_start:
-                text_parts.append(
-                    f"Current {search_label} streak: {duration_text}"
-                )
-                text_parts.append(
-                    f"Streak started: {self._format_relative_absolute_time(streak_start)}"
-                )
-            else:
-                text_parts.append(
-                    f"Current {search_label} streak: at least {duration_text}"
-                )
-                if streak_reaches_window_start:
-                    text_parts.append(
-                        f"The streak extends beyond the searched {hours}-hour window."
-                    )
-            text_parts = self._prepend_resolution_note(text_parts, resolution_note)
-            return {"content": [{"type": "text", "text": "\n".join(text_parts)}]}
-
-        if analysis == "stats":
-            stats_result = self._calculate_numeric_history_stats(
-                entity_states,
-                query_start_time,
-                query_end_time,
-            )
-
-            if stats_result is None:
-                return {
-                    "content": [
-                        {
-                            "type": "text",
-                            "text": (
-                                f"{friendly_name} ({history_entity_id})\n"
-                                "Recorder history did not contain numeric states in that window, so min/max/average could not be calculated."
-                            ),
-                        }
-                    ]
-                }
-
-            self.publish_progress(
-                "tool_complete",
-                "Recorder history numeric analysis complete",
-                tool="analyze_entity_history",
-                success=True,
-            )
-
-            min_time = stats_result["min_time"]
-            max_time = stats_result["max_time"]
-            text_parts = [
-                f"{friendly_name} ({history_entity_id})",
-                f"Current state: {current_state.state}",
-                f"Numeric recorder stats for the last {hours} hour{'s' if hours != 1 else ''}:",
-                f"Minimum: {self._format_number(stats_result['min'])}"
-                + (
-                    f" at {self._format_relative_absolute_time(min_time)}"
-                    if min_time is not None
-                    else ""
-                ),
-                f"Maximum: {self._format_number(stats_result['max'])}"
-                + (
-                    f" at {self._format_relative_absolute_time(max_time)}"
-                    if max_time is not None
-                    else ""
-                ),
-                f"Average: {self._format_number(stats_result['average'])} (time-weighted)",
-                f"Numeric state samples: {stats_result['sample_count']}",
-            ]
-            text_parts = self._prepend_resolution_note(text_parts, resolution_note)
-
-            return {"content": [{"type": "text", "text": "\n".join(text_parts)}]}
-
-        if target_filter_states:
-            matched_states = [
-                state
-                for state in entity_states
-                if state.state.casefold() in target_filter_states
-            ]
-        else:
-            matched_states = entity_states
-
-        if target_filter_states and not matched_states and len(history_candidates) > 1:
-            for candidate in history_candidates[1:]:
-                candidate_history = await _load_candidate_history(candidate)
-                if candidate_history is None:
-                    continue
-                candidate_matched_states = [
-                    state
-                    for state in candidate_history["entity_states"]
-                    if state.state.casefold() in candidate_history["target_filter_states"]
-                ]
-                if candidate_matched_states:
-                    history_entity_id = candidate_history["entity_id"]
-                    current_state = candidate_history["current_state"]
-                    friendly_name = candidate_history["friendly_name"]
-                    entity_states = candidate_history["entity_states"]
-                    target_filter_states = candidate_history["target_filter_states"]
-                    matched_states = candidate_matched_states
-                    resolution_note = candidate["note"]
-                    break
-
-        match_count = len(matched_states)
-
-        self.publish_progress(
-            "tool_complete",
-            "Recorder history analysis complete",
-            tool="analyze_entity_history",
-            success=True,
-            count=match_count,
-        )
-
-        search_label = self._describe_history_target(
-            args.get("state"), args.get("event")
-        )
-        noun = f"{search_label} event" if search_label != "change" else "state change"
-
-        text_parts = [
-            f"{friendly_name} ({history_entity_id})",
-            f"Current state: {current_state.state}",
-            f"Recorded {noun}{'s' if match_count != 1 else ''} in the last {hours} hour{'s' if hours != 1 else ''}: {match_count}",
-        ]
-
-        if target_filter_states:
-            text_parts.append(
-                f"Counted using recorder state{'s' if len(target_filter_states) != 1 else ''}: {', '.join(target_filter_states)}"
-            )
-
-        if analysis == "summary" and matched_states:
-            first_match = matched_states[0]
-            last_match = matched_states[-1]
-            first_when = first_match.last_changed or first_match.last_updated
-            last_when = last_match.last_changed or last_match.last_updated
-            text_parts.append(
-                f"First matching event in window: {self._format_relative_absolute_time(first_when)}"
-            )
-            text_parts.append(
-                f"Most recent matching event: {self._format_relative_absolute_time(last_when)}"
-            )
-
-        if analysis == "summary" and not matched_states:
-            text_parts.append("No matching recorder events were found in that window.")
-
-        text_parts = self._prepend_resolution_note(text_parts, resolution_note)
-        return {"content": [{"type": "text", "text": "\n".join(text_parts)}]}
-
-    async def tool_get_entity_state_at_time(self, args: Dict[str, Any]) -> Dict[str, Any]:
-        """Return an entity's recorder state at a specific point in time."""
-        entity_id = args.get("entity_id")
-        raw_datetime = args.get("datetime")
-        current_state = self.hass.states.get(entity_id)
-
-        if not current_state:
-            return {
-                "content": [
-                    {"type": "text", "text": f"Entity '{entity_id}' not found."}
-                ]
-            }
-
-        target_time = self._parse_history_datetime(raw_datetime)
-        if target_time is None:
-            return {
-                "content": [
-                    {
-                        "type": "text",
-                        "text": "Invalid datetime. Use an ISO 8601 timestamp or a Home Assistant-local datetime string.",
-                    }
-                ]
-            }
-
-        friendly_name = current_state.attributes.get("friendly_name", entity_id)
-        lookup_start = target_time + timedelta(microseconds=1)
-
-        self.publish_progress(
-            "tool_start",
-            f"Looking up recorder state for {entity_id} at {raw_datetime}",
-            tool="get_entity_state_at_time",
-            entity_id=entity_id,
-        )
-
-        try:
-            entity_states = await self._fetch_entity_history_states(
-                entity_id,
-                start_time=lookup_start,
-                end_time=lookup_start + timedelta(seconds=1),
-                descending=False,
-                include_start_time_state=True,
-            )
-        except Exception as err:
-            _LOGGER.error("Failed to look up recorder state for %s: %s", entity_id, err)
-            return {
-                "content": [
-                    {
-                        "type": "text",
-                        "text": f"Failed to retrieve recorder state: {err}",
-                    }
-                ]
-            }
-
-        state_at_time = entity_states[0] if entity_states else None
-        if state_at_time is not None:
-            state_when = state_at_time.last_changed or state_at_time.last_updated
-            if state_when > target_time:
-                state_at_time = None
-
-        self.publish_progress(
-            "tool_complete",
-            "Recorder point-in-time lookup complete",
-            tool="get_entity_state_at_time",
-            success=True,
-            found=state_at_time is not None,
-        )
-
-        target_local = self._format_absolute_time(target_time)
-        if state_at_time is None:
-            return {
-                "content": [
-                    {
-                        "type": "text",
-                        "text": (
-                            f"{friendly_name} ({entity_id})\n"
-                            f"No recorder state was available for {target_local}."
-                        ),
-                    }
-                ]
-            }
-
-        return {
-            "content": [
-                {
-                    "type": "text",
-                    "text": (
-                        f"{friendly_name} ({entity_id})\n"
-                        f"State at {target_local}: {state_at_time.state}\n"
-                        f"Current state: {current_state.state}"
-                    ),
-                }
-            ]
-        }
 
     async def tool_remember_memory(self, args: Dict[str, Any]) -> Dict[str, Any]:
         """Store a persisted memory with TTL."""
@@ -5720,608 +4123,6 @@ class MCPServer:
             "deleted": deleted,
         }
 
-    def _extract_history_states(
-        self, history_result: Dict[str, Any], entity_id: str
-    ) -> List[Any]:
-        """Extract an entity's state list from recorder helper results."""
-        return (
-            history_result.get(entity_id)
-            or history_result.get(entity_id.casefold())
-            or next(iter(history_result.values()), [])
-        )
-
-    def _normalize_history_targets(
-        self, raw_state: Any, raw_event: Any
-    ) -> List[str]:
-        """Normalize target states or semantic events into recorder states."""
-        values: List[str] = []
-
-        if isinstance(raw_state, list):
-            values.extend(str(value) for value in raw_state)
-        elif raw_state is not None:
-            values.append(str(raw_state))
-
-        if raw_event is not None:
-            values.append(str(raw_event))
-
-        expanded: List[str] = []
-        for value in values:
-            normalized = value.strip().casefold().replace(" ", "_")
-            if not normalized:
-                continue
-            expanded.extend(self._history_state_aliases(normalized))
-
-        deduped: List[str] = []
-        seen = set()
-        for value in expanded:
-            if value not in seen:
-                seen.add(value)
-                deduped.append(value)
-        return deduped
-
-    def _normalize_history_request_values(
-        self, raw_state: Any, raw_event: Any
-    ) -> List[str]:
-        """Normalize raw requested history values before alias expansion."""
-        values: List[str] = []
-
-        if isinstance(raw_state, list):
-            values.extend(str(value) for value in raw_state)
-        elif raw_state is not None:
-            values.append(str(raw_state))
-
-        if raw_event is not None:
-            values.append(str(raw_event))
-
-        normalized: List[str] = []
-        seen = set()
-        for value in values:
-            item = value.strip().casefold().replace(" ", "_")
-            if item and item not in seen:
-                seen.add(item)
-                normalized.append(item)
-        return normalized
-
-    def _choose_history_count_states(
-        self, entity_states: List[Any], raw_state: Any, raw_event: Any
-    ) -> List[str]:
-        """Choose the best recorder states to count for semantic event requests."""
-        present_states = {state.state.casefold() for state in entity_states}
-        requested_values = self._normalize_history_request_values(raw_state, raw_event)
-        chosen: List[str] = []
-        seen = set()
-
-        for value in requested_values:
-            aliases = self._history_state_aliases(value)
-            selected = next((alias for alias in aliases if alias in present_states), aliases[0])
-            if selected not in seen:
-                seen.add(selected)
-                chosen.append(selected)
-
-        return chosen or self._normalize_history_targets(raw_state, raw_event)
-
-    def _calculate_state_duration(
-        self,
-        entity_states: List[Any],
-        target_states: List[str],
-        start_time,
-        end_time,
-    ) -> Dict[str, Any]:
-        """Calculate total time spent in the target states over a window."""
-        total_duration = timedelta(0)
-        interval_count = 0
-        first_start = None
-        last_end = None
-
-        for index, state in enumerate(entity_states):
-            state_start = self._clip_time_to_window(
-                state.last_changed or state.last_updated,
-                start_time,
-                end_time,
-            )
-
-            if index + 1 < len(entity_states):
-                next_state = entity_states[index + 1]
-                state_end = self._clip_time_to_window(
-                    next_state.last_changed or next_state.last_updated,
-                    start_time,
-                    end_time,
-                )
-            else:
-                state_end = end_time
-
-            if state_end <= state_start:
-                continue
-
-            if state.state.casefold() not in target_states:
-                continue
-
-            total_duration += state_end - state_start
-            interval_count += 1
-            if first_start is None:
-                first_start = state_start
-            last_end = state_end
-
-        return {
-            "total_duration": total_duration,
-            "interval_count": interval_count,
-            "first_start": first_start,
-            "last_end": last_end,
-        }
-
-    def _calculate_numeric_history_stats(
-        self,
-        entity_states: List[Any],
-        start_time,
-        end_time,
-    ) -> Dict[str, Any] | None:
-        """Calculate numeric min/max/average across recorder states."""
-        numeric_points: List[tuple[float, Any]] = []
-        weighted_sum = 0.0
-        weighted_seconds = 0.0
-
-        for index, state in enumerate(entity_states):
-            numeric_value = self._coerce_numeric_state(state.state)
-            if numeric_value is None:
-                continue
-
-            state_time = self._clip_time_to_window(
-                state.last_changed or state.last_updated,
-                start_time,
-                end_time,
-            )
-            numeric_points.append((numeric_value, state_time))
-
-            if index + 1 < len(entity_states):
-                next_state = entity_states[index + 1]
-                state_end = self._clip_time_to_window(
-                    next_state.last_changed or next_state.last_updated,
-                    start_time,
-                    end_time,
-                )
-            else:
-                state_end = end_time
-
-            seconds = max((state_end - state_time).total_seconds(), 0)
-            if seconds > 0:
-                weighted_sum += numeric_value * seconds
-                weighted_seconds += seconds
-
-        if not numeric_points:
-            return None
-
-        min_value, min_time = min(numeric_points, key=lambda item: item[0])
-        max_value, max_time = max(numeric_points, key=lambda item: item[0])
-
-        if weighted_seconds > 0:
-            average = weighted_sum / weighted_seconds
-        else:
-            average = sum(value for value, _ in numeric_points) / len(numeric_points)
-
-        return {
-            "min": min_value,
-            "min_time": min_time,
-            "max": max_value,
-            "max_time": max_time,
-            "average": average,
-            "sample_count": len(numeric_points),
-        }
-
-    def _clip_time_to_window(self, when, start_time, end_time):
-        """Clip a timestamp to the requested analysis window."""
-        if when < start_time:
-            return start_time
-        if when > end_time:
-            return end_time
-        return when
-
-    def _coerce_numeric_state(self, value: Any) -> float | None:
-        """Convert a recorder state value into a numeric value when possible."""
-        try:
-            number = float(str(value).strip())
-        except (TypeError, ValueError):
-            return None
-
-        if number != number or number in (float("inf"), float("-inf")):
-            return None
-
-        return number
-
-    def _format_duration(self, duration: timedelta) -> str:
-        """Format a duration in human-friendly units."""
-        total_seconds = max(int(duration.total_seconds()), 0)
-        days, remainder = divmod(total_seconds, 86400)
-        hours, remainder = divmod(remainder, 3600)
-        minutes, seconds = divmod(remainder, 60)
-
-        parts: List[str] = []
-        if days:
-            parts.append(f"{days} day{'s' if days != 1 else ''}")
-        if hours:
-            parts.append(f"{hours} hour{'s' if hours != 1 else ''}")
-        if minutes:
-            parts.append(f"{minutes} minute{'s' if minutes != 1 else ''}")
-        if seconds and not parts:
-            parts.append(f"{seconds} second{'s' if seconds != 1 else ''}")
-
-        return ", ".join(parts) if parts else "0 seconds"
-
-    def _format_number(self, value: float) -> str:
-        """Format numeric values compactly for responses."""
-        if float(value).is_integer():
-            return str(int(value))
-        return f"{value:.3f}".rstrip("0").rstrip(".")
-
-    def _history_preferred_domains(self, requested_values: List[str]) -> Dict[str, int]:
-        """Return domain preferences for semantic history requests."""
-        domain_scores: Dict[str, int] = {}
-        for value in requested_values:
-            if value in {"locked", "unlocked"}:
-                domain_scores["lock"] = max(domain_scores.get("lock", 0), 120)
-            elif value in {"open", "opened", "opening", "close", "closed", "closing"}:
-                domain_scores["cover"] = max(domain_scores.get("cover", 0), 90)
-                domain_scores["binary_sensor"] = max(
-                    domain_scores.get("binary_sensor", 0), 80
-                )
-                domain_scores["lock"] = max(domain_scores.get("lock", 0), 40)
-            elif value in {"home", "away", "not_home"}:
-                domain_scores["person"] = max(domain_scores.get("person", 0), 100)
-                domain_scores["device_tracker"] = max(
-                    domain_scores.get("device_tracker", 0), 95
-                )
-            elif value in {"detected", "triggered", "clear", "cleared"}:
-                domain_scores["binary_sensor"] = max(
-                    domain_scores.get("binary_sensor", 0), 90
-                )
-        return domain_scores
-
-    def _history_preferred_device_classes(self, requested_values: List[str]) -> set[str]:
-        """Return preferred device classes for semantic history requests."""
-        device_classes: set[str] = set()
-        for value in requested_values:
-            if value in {"open", "opened", "opening", "close", "closed", "closing"}:
-                device_classes.update({"door", "window", "opening", "garage_door"})
-            elif value in {"detected", "triggered", "clear", "cleared"}:
-                device_classes.update(
-                    {
-                        "motion",
-                        "occupancy",
-                        "presence",
-                        "door",
-                        "window",
-                        "opening",
-                        "garage_door",
-                        "sound",
-                        "vibration",
-                    }
-                )
-        return device_classes
-
-    def _score_entity_for_history_request(
-        self,
-        state_obj: Any,
-        requested_values: List[str],
-        target_states: List[str],
-    ) -> int:
-        """Score how appropriate an entity is for a semantic history request."""
-        if state_obj is None or not requested_values:
-            return 0
-
-        domain = state_obj.entity_id.split(".", 1)[0]
-        device_class = str(state_obj.attributes.get("device_class", "")).casefold()
-        current_state = state_obj.state.casefold()
-        preferred_domains = self._history_preferred_domains(requested_values)
-        preferred_device_classes = self._history_preferred_device_classes(
-            requested_values
-        )
-
-        score = preferred_domains.get(domain, 0)
-        if device_class and device_class in preferred_device_classes:
-            score += 25
-        if current_state in target_states:
-            score += 20
-
-        return score
-
-    def _history_relation_tokens(self, value: Any) -> List[str]:
-        """Normalize an entity-related name into meaningful comparison tokens."""
-        if value in (None, ""):
-            return []
-
-        stopwords = {
-            "a",
-            "an",
-            "the",
-            "entity",
-            "device",
-            "binary",
-            "sensor",
-            "contact",
-            "contacts",
-            "lock",
-            "locks",
-            "locked",
-            "unlock",
-            "unlocked",
-            "deadbolt",
-            "bolt",
-        }
-        raw_tokens = re.findall(r"[a-z0-9]+", str(value).casefold())
-        filtered = [
-            token
-            for token in raw_tokens
-            if len(token) > 1 and token not in stopwords
-        ]
-        return filtered or [token for token in raw_tokens if len(token) > 1]
-
-    def _history_relation_texts(self, state_obj: Any) -> List[str]:
-        """Collect stable entity name variants for related-entity matching."""
-        if state_obj is None:
-            return []
-
-        texts: List[str] = []
-        for value in (
-            getattr(state_obj, "name", None),
-            state_obj.attributes.get("friendly_name"),
-            state_obj.entity_id.split(".", 1)[1].replace("_", " "),
-        ):
-            text = str(value).strip() if value not in (None, "") else ""
-            if text and text not in texts:
-                texts.append(text)
-        return texts
-
-    def _history_entity_location_signature(
-        self, entity_id: str
-    ) -> Tuple[str | None, str | None, str | None]:
-        """Return device, area, and floor identifiers for an entity."""
-        entity_registry = er.async_get(self.hass)
-        device_registry = dr.async_get(self.hass)
-        area_registry = ar.async_get(self.hass)
-        floor_registry = fr.async_get(self.hass) if fr else None
-
-        entity_entry = entity_registry.async_get(entity_id)
-        if entity_entry is None:
-            return None, None, None
-
-        device_entry = (
-            device_registry.async_get(entity_entry.device_id)
-            if entity_entry.device_id
-            else None
-        )
-        area_id = entity_entry.area_id or (device_entry.area_id if device_entry else None)
-        area_entry = area_registry.async_get_area(area_id) if area_id else None
-        floor_id = getattr(area_entry, "floor_id", None) if area_entry else None
-        if floor_id and floor_registry is not None:
-            floor_entry = floor_registry.async_get_floor(floor_id)
-            floor_id = floor_entry.floor_id if floor_entry else floor_id
-
-        return entity_entry.device_id, area_id, floor_id
-
-    def _score_history_entity_relatedness(
-        self, current_state: Any, candidate_state: Any
-    ) -> Tuple[int, bool]:
-        """Score how likely two entities describe the same real-world thing."""
-        if current_state is None or candidate_state is None:
-            return 0, False
-
-        current_device_id, current_area_id, current_floor_id = (
-            self._history_entity_location_signature(current_state.entity_id)
-        )
-        candidate_device_id, candidate_area_id, candidate_floor_id = (
-            self._history_entity_location_signature(candidate_state.entity_id)
-        )
-
-        best_name_score = 0
-        for current_text in self._history_relation_texts(current_state):
-            current_tokens = self._history_relation_tokens(current_text)
-            current_canonical = " ".join(current_tokens)
-            if not current_tokens:
-                continue
-
-            for candidate_text in self._history_relation_texts(candidate_state):
-                candidate_tokens = self._history_relation_tokens(candidate_text)
-                candidate_canonical = " ".join(candidate_tokens)
-                if not candidate_tokens:
-                    continue
-
-                overlap = set(current_tokens) & set(candidate_tokens)
-                if not overlap:
-                    continue
-
-                score = len(overlap) * 22
-                coverage = len(overlap) / min(len(current_tokens), len(candidate_tokens))
-                score += int(coverage * 55)
-
-                if current_canonical and current_canonical == candidate_canonical:
-                    score += 80
-                elif (
-                    current_canonical
-                    and candidate_canonical
-                    and min(len(current_canonical), len(candidate_canonical)) >= 5
-                    and (
-                        current_canonical in candidate_canonical
-                        or candidate_canonical in current_canonical
-                    )
-                ):
-                    score += 40
-
-                best_name_score = max(best_name_score, score)
-
-        same_device = bool(
-            current_device_id
-            and candidate_device_id
-            and current_device_id == candidate_device_id
-        )
-        if same_device:
-            best_name_score += 120
-        elif best_name_score:
-            if current_area_id and candidate_area_id and current_area_id == candidate_area_id:
-                best_name_score += 10
-            if (
-                current_floor_id
-                and candidate_floor_id
-                and current_floor_id == candidate_floor_id
-            ):
-                best_name_score += 5
-
-        return best_name_score, same_device
-
-    def _resolve_history_entity_for_request(
-        self,
-        entity_id: str,
-        raw_state: Any,
-        raw_event: Any,
-    ) -> tuple[str, str | None]:
-        """Resolve the best entity for semantic history requests when needed."""
-        candidates = self._history_resolution_candidates(entity_id, raw_state, raw_event)
-        best_candidate = candidates[0]
-        return best_candidate["entity_id"], best_candidate["note"]
-
-    def _history_resolution_candidates(
-        self,
-        entity_id: str,
-        raw_state: Any,
-        raw_event: Any,
-        *,
-        max_candidates: int = 4,
-    ) -> List[Dict[str, Any]]:
-        """Return strong related-entity candidates for a semantic history request."""
-        requested_values = self._normalize_history_request_values(raw_state, raw_event)
-        if not requested_values:
-            return [{"entity_id": entity_id, "note": None}]
-
-        target_states = self._normalize_history_targets(raw_state, raw_event)
-        current_state = self.hass.states.get(entity_id)
-        if current_state is None:
-            return [{"entity_id": entity_id, "note": None}]
-
-        current_score = self._score_entity_for_history_request(
-            current_state,
-            requested_values,
-            target_states,
-        )
-        candidates: List[Dict[str, Any]] = [
-            {
-                "entity_id": entity_id,
-                "note": None,
-                "priority": current_score + 150,
-                "semantic_score": current_score,
-                "relation_score": 0,
-            }
-        ]
-
-        requested_label = self._describe_history_target(raw_state, raw_event)
-
-        for sibling_state in self.hass.states.async_all():
-            sibling_entity_id = sibling_state.entity_id
-            if sibling_entity_id == entity_id:
-                continue
-            if not async_should_expose(self.hass, "conversation", sibling_entity_id):
-                continue
-            sibling_score = self._score_entity_for_history_request(
-                sibling_state,
-                requested_values,
-                target_states,
-            )
-            if sibling_score <= 0:
-                continue
-
-            relation_score, same_device = self._score_history_entity_relatedness(
-                current_state,
-                sibling_state,
-            )
-            if relation_score <= 0:
-                continue
-
-            total_score = sibling_score + relation_score
-            if same_device:
-                if (
-                    sibling_score <= current_score
-                    or total_score < current_score + 40
-                ):
-                    continue
-            else:
-                if (
-                    sibling_score < current_score + 20
-                    or relation_score < 90
-                    or total_score < current_score + 100
-                ):
-                    continue
-
-            reason_text = (
-                "is on the same device and"
-                if same_device
-                else "strongly matches the same named thing and"
-            )
-            candidates.append(
-                {
-                    "entity_id": sibling_entity_id,
-                    "note": (
-                        f"Using related entity {sibling_entity_id} because it {reason_text} "
-                        f"the requested {requested_label} history applies more directly to that entity "
-                        f"than {entity_id}."
-                    ),
-                    "priority": total_score + (40 if same_device else 0),
-                    "semantic_score": sibling_score,
-                    "relation_score": relation_score,
-                }
-            )
-
-        candidates.sort(
-            key=lambda item: (
-                -int(item["priority"]),
-                -int(item["semantic_score"]),
-                item["entity_id"],
-            )
-        )
-        return candidates[:max_candidates]
-
-    def _prepend_resolution_note(
-        self, text_parts: List[str], resolution_note: str | None
-    ) -> List[str]:
-        """Prepend a history-entity resolution note when a sibling entity was used."""
-        if not resolution_note:
-            return text_parts
-        return [resolution_note, ""] + text_parts
-
-    def _history_state_aliases(self, value: str) -> List[str]:
-        """Map semantic event words to likely recorder states."""
-        aliases = {
-            "open": ["open", "opening", "on"],
-            "opened": ["open", "opening", "on"],
-            "opening": ["opening", "open", "on"],
-            "close": ["closed", "closing", "off"],
-            "closed": ["closed", "closing", "off"],
-            "closing": ["closing", "closed", "off"],
-            "on": ["on"],
-            "turned_on": ["on"],
-            "enabled": ["on"],
-            "active": ["on"],
-            "off": ["off"],
-            "turned_off": ["off"],
-            "disabled": ["off"],
-            "locked": ["locked"],
-            "unlocked": ["unlocked"],
-            "detected": ["detected", "on", "open"],
-            "triggered": ["triggered", "on"],
-            "clear": ["clear", "off", "closed"],
-            "cleared": ["clear", "off", "closed"],
-            "home": ["home"],
-            "away": ["away", "not_home"],
-            "not_home": ["not_home", "away"],
-        }
-        return aliases.get(value, [value])
-
-    def _describe_history_target(self, raw_state: Any, raw_event: Any) -> str:
-        """Create a readable description of the recorder search target."""
-        if raw_event:
-            return str(raw_event).replace("_", " ").strip()
-        if isinstance(raw_state, list):
-            return " / ".join(str(value).replace("_", " ").strip() for value in raw_state)
-        if raw_state:
-            return str(raw_state).replace("_", " ").strip()
-        return "change"
-
     def _format_relative_time(self, when) -> str:
         """Format a timestamp relative to now."""
         now = dt_util.utcnow()
@@ -6367,22 +4168,6 @@ class MCPServer:
         relative = self._format_relative_time(when)
         absolute = self._format_absolute_time(when)
         return f"{relative} at {absolute}"
-
-    def _parse_history_datetime(self, value: Any):
-        """Parse a recorder lookup datetime and assume local time if naive."""
-        if value is None:
-            return None
-
-        parsed = dt_util.parse_datetime(str(value))
-        if parsed is None:
-            return None
-
-        if parsed.tzinfo is None:
-            parsed = parsed.replace(
-                tzinfo=getattr(dt_util, "DEFAULT_TIME_ZONE", dt_util.now().tzinfo)
-            )
-
-        return dt_util.as_utc(parsed)
 
     def _coerce_int_arg(
         self, value: Any, *, default: int, minimum: int, maximum: int
@@ -6511,53 +4296,6 @@ class MCPServer:
 
         return lines
 
-    def _build_history_search_windows(self, max_hours: int) -> List[int]:
-        """Build progressively larger recorder search windows."""
-        windows = [24, 168, 720, max_hours]
-        deduped: List[int] = []
-        seen = set()
-
-        for window in windows:
-            clamped = min(window, max_hours)
-            if clamped > 0 and clamped not in seen:
-                seen.add(clamped)
-                deduped.append(clamped)
-
-        return deduped
-
-    async def _fetch_entity_history_states(
-        self,
-        entity_id: str,
-        hours: int | None = None,
-        *,
-        start_time=None,
-        end_time=None,
-        descending: bool = True,
-        limit: int | None = None,
-        include_start_time_state: bool = False,
-    ) -> List[Any]:
-        """Fetch recorder-backed history states for a single entity."""
-        query_end_time = end_time or dt_util.utcnow()
-        query_start_time = start_time
-        if query_start_time is None:
-            if hours is None:
-                raise ValueError("Either hours or start_time must be provided")
-            query_start_time = query_end_time - timedelta(hours=hours)
-
-        states = await self.hass.async_add_executor_job(
-            lambda: history.state_changes_during_period(
-                self.hass,
-                query_start_time,
-                end_time=query_end_time,
-                entity_id=entity_id,
-                no_attributes=True,
-                descending=descending,
-                limit=limit,
-                include_start_time_state=include_start_time_state,
-            )
-        )
-        return self._extract_history_states(states, entity_id)
-
     def _friendly_names_for_entities(self, entity_ids: List[str]) -> List[str]:
         """Resolve entity IDs to friendly names."""
         names = []
@@ -6591,1077 +4329,6 @@ class MCPServer:
         else:
             _LOGGER.warning(f"Service validation failed: {result}")
             raise ValueError(result)  # Returns error message
-
-    def _prepare_response_service_data(
-        self,
-        domain: str,
-        service: str,
-        data: Dict[str, Any] | None,
-        *,
-        resolved_target: Dict[str, Any] | None = None,
-    ) -> Dict[str, Any]:
-        """Normalize service-response data and fill safe defaults."""
-        prepared = dict(data or {})
-
-        if (
-            domain == "weather"
-            and service in {"get_forecast", "get_forecasts"}
-        ):
-            requested_type = str(prepared.get("type") or "").strip().lower() or None
-            supported_types = self._get_weather_target_forecast_types(
-                resolved_target
-            )
-
-            chosen_type = requested_type
-            if supported_types:
-                if requested_type not in supported_types:
-                    chosen_type = self._select_preferred_weather_forecast_type(
-                        supported_types
-                    )
-                    if requested_type and chosen_type and chosen_type != requested_type:
-                        entity_ids = self._normalize_target_values(
-                            (resolved_target or {}).get("entity_id")
-                        )
-                        _LOGGER.info(
-                            "Weather forecast type '%s' is not supported by target %s; "
-                            "using '%s' instead from supported types %s",
-                            requested_type,
-                            entity_ids or "entities",
-                            chosen_type,
-                            supported_types,
-                        )
-
-            if chosen_type:
-                prepared["type"] = chosen_type
-
-        return prepared
-
-    async def _resolve_weather_forecast_target(
-        self,
-        *,
-        entity_id: str | None = None,
-        area: str | None = None,
-        floor: str | None = None,
-        label: str | None = None,
-        name_contains: str | None = None,
-    ) -> tuple[Dict[str, Any], Dict[str, str]]:
-        """Resolve the best weather entity target for a forecast request."""
-        if entity_id:
-            resolved_target = await self.resolve_target({"entity_id": entity_id})
-            resolved_target = self._restrict_resolved_target_to_domain(
-                resolved_target, "weather"
-            )
-            selected_entity_id = self._normalize_target_values(
-                resolved_target.get("entity_id")
-            )[0]
-            state = self.hass.states.get(selected_entity_id)
-            friendly_name = (
-                state.attributes.get("friendly_name")
-                if state and state.attributes.get("friendly_name")
-                else selected_entity_id
-            )
-            return resolved_target, {
-                "entity_id": selected_entity_id,
-                "name": str(friendly_name),
-            }
-
-        candidates = await self.discovery.discover_entities(
-            domain="weather",
-            area=area,
-            floor=floor,
-            label=label,
-            name_contains=name_contains,
-            limit=10,
-        )
-        if not candidates:
-            raise ValueError(
-                "No exposed weather entity was found in Home Assistant. "
-                "Try discover_entities(domain='weather') to inspect available weather entities."
-            )
-
-        selected = sorted(
-            candidates,
-            key=self._weather_entity_candidate_sort_key,
-        )[0]
-        return {"entity_id": [selected["entity_id"]]}, {
-            "entity_id": str(selected["entity_id"]),
-            "name": str(selected.get("name") or selected["entity_id"]),
-        }
-
-    async def _resolve_calendar_event_targets(
-        self,
-        *,
-        entity_id: str | None = None,
-        area: str | None = None,
-        floor: str | None = None,
-        label: str | None = None,
-        query: str | None = None,
-        name_contains: str | None = None,
-    ) -> tuple[Dict[str, Any], list[Dict[str, str]], bool]:
-        """Resolve one or more calendar targets for a search."""
-        if entity_id:
-            resolved_target = await self.resolve_target({"entity_id": entity_id})
-            resolved_target = self._restrict_resolved_target_to_domain(
-                resolved_target, "calendar"
-            )
-            entity_ids = self._normalize_target_values(resolved_target.get("entity_id"))
-            selected = [
-                {
-                    "entity_id": candidate_entity_id,
-                    "name": self._friendly_entity_name(candidate_entity_id),
-                }
-                for candidate_entity_id in entity_ids
-            ]
-            return resolved_target, selected, False
-
-        explicit_name_filter = bool(name_contains)
-        candidates: list[Dict[str, Any]] = []
-        name_filters = self._build_calendar_name_filters(
-            query=query,
-            name_contains=name_contains,
-        )
-        for discovery_name in name_filters:
-            candidates = await self.discovery.discover_entities(
-                domain="calendar",
-                area=area,
-                floor=floor,
-                label=label,
-                name_contains=discovery_name,
-                limit=25,
-            )
-            if candidates:
-                break
-        fallback_used = False
-        if not candidates and query and not explicit_name_filter:
-            candidates = await self.discovery.discover_entities(
-                domain="calendar",
-                area=area,
-                floor=floor,
-                label=label,
-                limit=25,
-            )
-            fallback_used = True
-
-        if not candidates:
-            if explicit_name_filter:
-                raise ValueError(
-                    "No exposed calendar matched that name. "
-                    "Try discover_entities(domain='calendar') to inspect available calendars."
-                )
-            raise ValueError(
-                "No exposed calendar was found in Home Assistant. "
-                "Try discover_entities(domain='calendar') to inspect available calendars."
-            )
-
-        selected = [
-            {
-                "entity_id": str(candidate["entity_id"]),
-                "name": str(candidate.get("name") or candidate["entity_id"]),
-            }
-            for candidate in candidates
-        ]
-        selected.sort(key=lambda item: item["name"].casefold())
-        return {"entity_id": [item["entity_id"] for item in selected]}, selected, fallback_used
-
-    def _build_calendar_name_filters(
-        self,
-        *,
-        query: str | None,
-        name_contains: str | None,
-    ) -> list[str | None]:
-        """Build progressively broader calendar-name filters before event-text fallback."""
-        explicit_name = str(name_contains or "").strip()
-        if explicit_name:
-            return [explicit_name]
-
-        raw_query = str(query or "").strip()
-        if not raw_query:
-            return [None]
-
-        filters: list[str | None] = [raw_query]
-        generic_words = {
-            "calendar",
-            "event",
-            "events",
-            "game",
-            "games",
-            "match",
-            "matches",
-            "schedule",
-            "schedules",
-            "next",
-            "upcoming",
-        }
-        query_tokens = re.findall(r"[a-z0-9']+", raw_query.casefold())
-        simplified_tokens = [token for token in query_tokens if token not in generic_words]
-        simplified_query = " ".join(simplified_tokens).strip()
-        if simplified_query and simplified_query != raw_query.casefold():
-            filters.append(simplified_query)
-
-        return filters
-
-    @staticmethod
-    def _weather_entity_candidate_sort_key(entity: Dict[str, Any]) -> Tuple[int, int, str, str]:
-        """Rank weather entity candidates with stable generic preferences."""
-        name = str(entity.get("name") or "").casefold()
-        entity_id = str(entity.get("entity_id") or "").casefold()
-        service_hint = 0 if entity.get("forecast_service_supported") else 1
-        attribute_hint = 0 if entity.get("forecast_available") else 1
-        return (service_hint, attribute_hint, name or entity_id, entity_id)
-
-    def _parse_weather_forecast_when(
-        self, when_value: str | None
-    ) -> tuple[date | None, str, str | None]:
-        """Parse a weather forecast day request."""
-        now_local = dt_util.as_local(dt_util.now())
-        when_text = str(when_value or "tomorrow").strip().casefold()
-        if not when_text or when_text == "tomorrow":
-            return now_local.date() + timedelta(days=1), "tomorrow", None
-        if when_text == "today":
-            return now_local.date(), "today", None
-
-        parsed_date = dt_util.parse_date(when_text)
-        if parsed_date is not None:
-            return parsed_date, self._format_weather_day_label(parsed_date), None
-
-        parsed_datetime = dt_util.parse_datetime(when_text)
-        if parsed_datetime is not None:
-            if parsed_datetime.tzinfo is None:
-                parsed_datetime = parsed_datetime.replace(
-                    tzinfo=getattr(dt_util, "DEFAULT_TIME_ZONE", now_local.tzinfo)
-                )
-            local_when = dt_util.as_local(parsed_datetime)
-            return local_when.date(), self._format_weather_day_label(local_when.date()), None
-
-        return None, when_text, (
-            "Unable to parse 'when'. Use 'today', 'tomorrow', or a local date like '2026-04-13'."
-        )
-
-    def _format_weather_day_label(self, target_date: date) -> str:
-        """Format a forecast day label for user-facing summaries."""
-        today = dt_util.as_local(dt_util.now()).date()
-        if target_date == today:
-            return "today"
-        if target_date == today + timedelta(days=1):
-            return "tomorrow"
-        return target_date.strftime("%A, %b %d").replace(" 0", " ")
-
-    def _extract_weather_entries_for_date(
-        self,
-        forecast_entries: List[Dict[str, Any]],
-        *,
-        target_date: date,
-        forecast_type: str | None,
-    ) -> List[Dict[str, Any]]:
-        """Return forecast entries that map to the requested local date."""
-        matching: List[tuple[datetime | None, Dict[str, Any]]] = []
-        for entry in forecast_entries:
-            entry_date = self._parse_weather_forecast_entry_date(entry)
-            if entry_date and entry_date == target_date:
-                matching.append((self._parse_weather_forecast_entry_datetime(entry), entry))
-
-        if matching:
-            matching.sort(key=lambda item: item[0] or datetime.min.replace(tzinfo=dt_util.UTC))
-            return [entry for _, entry in matching]
-
-        today = dt_util.as_local(dt_util.now()).date()
-        day_offset = (target_date - today).days
-        if day_offset < 0:
-            return []
-
-        if forecast_type == "daily" and day_offset < len(forecast_entries):
-            return [forecast_entries[day_offset]]
-
-        if forecast_type == "twice_daily":
-            start = day_offset * 2
-            if start < len(forecast_entries):
-                return forecast_entries[start : start + 2]
-
-        if forecast_type == "hourly":
-            start = day_offset * 24
-            if start < len(forecast_entries):
-                return forecast_entries[start : start + 24]
-
-        return []
-
-    def _parse_weather_forecast_entry_datetime(
-        self, entry: Dict[str, Any]
-    ) -> datetime | None:
-        """Parse a weather forecast entry datetime to local time."""
-        raw_value = entry.get("datetime")
-        if raw_value is None:
-            raw_value = entry.get("date")
-        if raw_value is None:
-            return None
-
-        parsed_datetime = dt_util.parse_datetime(str(raw_value))
-        if parsed_datetime is not None:
-            if parsed_datetime.tzinfo is None:
-                parsed_datetime = parsed_datetime.replace(
-                    tzinfo=getattr(dt_util, "DEFAULT_TIME_ZONE", dt_util.now().tzinfo)
-                )
-            return dt_util.as_local(parsed_datetime)
-
-        parsed_date = dt_util.parse_date(str(raw_value))
-        if parsed_date is None:
-            return None
-
-        local_tz = getattr(dt_util, "DEFAULT_TIME_ZONE", dt_util.now().tzinfo)
-        return datetime.combine(parsed_date, time.min, tzinfo=local_tz)
-
-    def _parse_weather_forecast_entry_date(
-        self, entry: Dict[str, Any]
-    ) -> date | None:
-        """Parse just the local date for a weather forecast entry."""
-        parsed_datetime = self._parse_weather_forecast_entry_datetime(entry)
-        if parsed_datetime is not None:
-            return parsed_datetime.date()
-        return None
-
-    def _summarize_requested_weather_forecast(
-        self,
-        *,
-        entity_name: str,
-        entity_id: str,
-        forecast_entries: List[Dict[str, Any]],
-        forecast_type: str | None,
-        target_date: date,
-        day_label: str,
-    ) -> str | None:
-        """Summarize a requested weather forecast day from returned forecast data."""
-        day_entries = self._extract_weather_entries_for_date(
-            forecast_entries,
-            target_date=target_date,
-            forecast_type=forecast_type,
-        )
-        if not day_entries:
-            return None
-
-        intro = f"{day_label.capitalize()} for {entity_name}:"
-        if forecast_type == "twice_daily":
-            parts = []
-            for entry in day_entries[:2]:
-                part_label = self._describe_weather_forecast_part(entry)
-                parts.append(f"{part_label}: {self._format_weather_forecast_entry(entry)}")
-            return f"{intro} {'; '.join(parts)}."
-
-        if forecast_type == "hourly" and len(day_entries) > 2:
-            conditions = [
-                str(entry.get("condition")).replace("_", " ")
-                for entry in day_entries
-                if entry.get("condition")
-            ]
-            temps = [
-                self._coerce_weather_temperature(entry.get("temperature"))
-                for entry in day_entries
-            ]
-            temps = [temp for temp in temps if temp is not None]
-            dominant_condition = None
-            if conditions:
-                dominant_condition = Counter(conditions).most_common(1)[0][0]
-
-            detail_parts = []
-            if dominant_condition:
-                detail_parts.append(dominant_condition)
-            if temps:
-                detail_parts.append(
-                    f"high {self._format_weather_temperature(max(temps))}, low {self._format_weather_temperature(min(temps))}"
-                )
-            preview_entries = [
-                f"{self._describe_weather_forecast_part(entry)} {self._format_weather_forecast_entry(entry)}"
-                for entry in day_entries[:3]
-            ]
-            if preview_entries:
-                detail_parts.append("early outlook: " + "; ".join(preview_entries))
-            return f"{intro} {'. '.join(detail_parts)}."
-
-        return f"{intro} {self._format_weather_forecast_entry(day_entries[0])}."
-
-    def _build_calendar_search_window(
-        self,
-        *,
-        when_value: str | None,
-        days: int,
-    ) -> tuple[datetime | None, datetime | None, str]:
-        """Build the UTC search window for a calendar query."""
-        local_tz = getattr(dt_util, "DEFAULT_TIME_ZONE", dt_util.now().tzinfo)
-        now_local = dt_util.as_local(dt_util.utcnow())
-        raw_when = (when_value or "").strip()
-        normalized = raw_when.casefold()
-
-        if not raw_when or normalized == "now":
-            window_start = dt_util.utcnow()
-            window_end = window_start + timedelta(days=days)
-            return window_start, window_end, f"the next {days} day{'s' if days != 1 else ''}"
-
-        if normalized == "today":
-            start_local = datetime.combine(now_local.date(), time.min, tzinfo=local_tz)
-            end_local = start_local + timedelta(days=1)
-            return dt_util.as_utc(start_local), dt_util.as_utc(end_local), "today"
-
-        if normalized == "tomorrow":
-            start_local = datetime.combine(
-                now_local.date() + timedelta(days=1),
-                time.min,
-                tzinfo=local_tz,
-            )
-            end_local = start_local + timedelta(days=1)
-            return dt_util.as_utc(start_local), dt_util.as_utc(end_local), "tomorrow"
-
-        parsed_date = dt_util.parse_date(raw_when)
-        if parsed_date is not None:
-            start_local = datetime.combine(parsed_date, time.min, tzinfo=local_tz)
-            end_local = start_local + timedelta(days=1)
-            return dt_util.as_utc(start_local), dt_util.as_utc(end_local), raw_when
-
-        parsed_datetime = dt_util.parse_datetime(raw_when)
-        if parsed_datetime is None:
-            return None, None, raw_when
-        if parsed_datetime.tzinfo is None:
-            parsed_datetime = parsed_datetime.replace(tzinfo=local_tz)
-        window_start = dt_util.as_utc(parsed_datetime)
-        window_end = window_start + timedelta(days=days)
-        return window_start, window_end, raw_when
-
-    def _collect_calendar_event_matches(
-        self,
-        *,
-        response: Dict[str, Any],
-        selected_calendars: list[Dict[str, str]],
-        event_text: str | None,
-    ) -> list[Dict[str, Any]]:
-        """Collect matching calendar events sorted by upcoming start time."""
-        selected_by_entity = {
-            item["entity_id"]: item["name"] for item in selected_calendars
-        }
-        matches: list[Dict[str, Any]] = []
-        for entity_id, payload in response.items():
-            events = payload.get("events") if isinstance(payload, dict) else None
-            if not isinstance(events, list):
-                continue
-            calendar_name = selected_by_entity.get(entity_id, entity_id)
-            for event in events:
-                if not isinstance(event, dict):
-                    continue
-                if event_text and not self._calendar_event_matches_text(event, event_text):
-                    continue
-                start_at = self._parse_calendar_event_start_datetime(event)
-                if start_at is None:
-                    continue
-                matches.append(
-                    {
-                        "calendar_entity_id": entity_id,
-                        "calendar_name": calendar_name,
-                        "summary": str(
-                            event.get("summary")
-                            or event.get("title")
-                            or "Untitled event"
-                        ),
-                        "description": str(event.get("description") or "").strip(),
-                        "location": str(event.get("location") or "").strip(),
-                        "start": self._serialize_service_response_value(
-                            self._extract_calendar_event_start_value(event)
-                        ),
-                        "end": self._serialize_service_response_value(event.get("end")),
-                        "all_day": self._calendar_event_is_all_day(event),
-                        "start_at": start_at,
-                    }
-                )
-        matches.sort(key=lambda item: item["start_at"])
-        return matches
-
-    def _summarize_calendar_matches(
-        self,
-        *,
-        matches: list[Dict[str, Any]],
-        query: str | None,
-        event_text: str | None,
-        window_label: str,
-    ) -> str:
-        """Summarize one or more calendar event matches."""
-        if len(matches) == 1:
-            event = matches[0]
-            qualifier = "matching calendar event" if (query or event_text) else "calendar event"
-            return (
-                f"Next {qualifier}: {self._describe_calendar_match(event)}."
-            )
-
-        lead = "Upcoming matching calendar events" if (query or event_text) else "Upcoming calendar events"
-        lines = [f"{lead} in {window_label}:"]
-        for event in matches:
-            lines.append(f"- {self._describe_calendar_match(event)}")
-        return "\n".join(lines)
-
-    def _describe_calendar_match(self, event: Dict[str, Any]) -> str:
-        """Describe one calendar event compactly."""
-        summary = str(event.get("summary") or "Untitled event")
-        calendar_name = str(event.get("calendar_name") or event.get("calendar_entity_id") or "")
-        when = self._format_calendar_event_when(event)
-        parts = [summary]
-        if calendar_name and calendar_name.casefold() not in summary.casefold():
-            parts.append(f"on {calendar_name}")
-        if when:
-            parts.append(when)
-        location = str(event.get("location") or "").strip()
-        if location:
-            parts.append(f"at {location}")
-        return ", ".join(parts)
-
-    def _format_calendar_event_when(self, event: Dict[str, Any]) -> str:
-        """Format a calendar event start in friendly upcoming language."""
-        start_at = event.get("start_at")
-        if not isinstance(start_at, datetime):
-            return ""
-        if event.get("all_day"):
-            return self._format_calendar_all_day(start_at)
-        return self._format_future_absolute_time(start_at)
-
-    def _format_calendar_all_day(self, when: datetime) -> str:
-        """Format an all-day calendar date."""
-        local_when = dt_util.as_local(when)
-        today_local = dt_util.as_local(dt_util.utcnow()).date()
-        day_delta = (local_when.date() - today_local).days
-        if day_delta == 0:
-            return "all day today"
-        if day_delta == 1:
-            return "all day tomorrow"
-        if day_delta > 1 and day_delta < 7:
-            return f"all day in {day_delta} days"
-        date_text = local_when.strftime("%b %d").replace(" 0", " ")
-        if local_when.year != today_local.year:
-            date_text += f", {local_when.year}"
-        return f"all day on {date_text}"
-
-    def _format_future_absolute_time(self, when: datetime) -> str:
-        """Format a future timestamp with relative and absolute context."""
-        now = dt_util.utcnow()
-        delta_seconds = (when - now).total_seconds()
-        absolute = self._format_absolute_time(when)
-        if delta_seconds <= 0:
-            return absolute
-        if delta_seconds < 3600:
-            minutes = max(1, int(delta_seconds / 60))
-            return f"in {minutes} minute{'s' if minutes != 1 else ''} at {absolute}"
-        if delta_seconds < 86400:
-            hours = max(1, int(delta_seconds / 3600))
-            return f"in {hours} hour{'s' if hours != 1 else ''} at {absolute}"
-        if delta_seconds < 604800:
-            days = max(1, int(delta_seconds / 86400))
-            return f"in {days} day{'s' if days != 1 else ''} at {absolute}"
-        weeks = max(1, int(delta_seconds / 604800))
-        return f"in {weeks} week{'s' if weeks != 1 else ''} at {absolute}"
-
-    def _calendar_event_matches_text(self, event: Dict[str, Any], text: str) -> bool:
-        """Return whether a calendar event matches free text."""
-        query = text.casefold().strip()
-        if not query:
-            return True
-        haystack = " ".join(
-            str(event.get(key) or "").casefold()
-            for key in ("summary", "title", "description", "location")
-        )
-        return query in haystack
-
-    def _calendar_event_is_all_day(self, event: Dict[str, Any]) -> bool:
-        """Return whether a calendar event is all-day."""
-        start_value = event.get("start")
-        if isinstance(start_value, dict):
-            return bool(start_value.get("date")) and not bool(
-                start_value.get("dateTime") or start_value.get("datetime")
-            )
-        if isinstance(start_value, str):
-            return bool(re.fullmatch(r"\d{4}-\d{2}-\d{2}", start_value.strip()))
-        return False
-
-    def _extract_calendar_event_start_value(self, event: Dict[str, Any]) -> Any:
-        """Extract the raw calendar event start field."""
-        start_value = event.get("start")
-        if isinstance(start_value, dict):
-            return (
-                start_value.get("dateTime")
-                or start_value.get("datetime")
-                or start_value.get("date")
-            )
-        return start_value
-
-    def _parse_calendar_event_start_datetime(self, event: Dict[str, Any]) -> datetime | None:
-        """Parse a calendar event start into UTC for sorting."""
-        start_value = self._extract_calendar_event_start_value(event)
-        local_tz = getattr(dt_util, "DEFAULT_TIME_ZONE", dt_util.now().tzinfo)
-        if self._calendar_event_is_all_day(event):
-            parsed_date = dt_util.parse_date(str(start_value))
-            if parsed_date is None:
-                return None
-            return dt_util.as_utc(
-                datetime.combine(parsed_date, time.min, tzinfo=local_tz)
-            )
-
-        parsed_datetime = dt_util.parse_datetime(str(start_value))
-        if parsed_datetime is None:
-            return None
-        if parsed_datetime.tzinfo is None:
-            parsed_datetime = parsed_datetime.replace(tzinfo=local_tz)
-        return dt_util.as_utc(parsed_datetime)
-
-    def _friendly_entity_name(self, entity_id: str) -> str:
-        """Return a friendly entity name when available."""
-        state = self.hass.states.get(entity_id)
-        if state is not None and state.attributes.get("friendly_name"):
-            return str(state.attributes["friendly_name"])
-        return entity_id
-
-    def _describe_weather_forecast_part(self, entry: Dict[str, Any]) -> str:
-        """Describe a weather forecast segment like morning/evening/time."""
-        parsed_datetime = self._parse_weather_forecast_entry_datetime(entry)
-        if parsed_datetime is not None:
-            hour = parsed_datetime.hour
-            if entry.get("is_daytime") is False or hour >= 18:
-                return "evening"
-            if 5 <= hour < 12:
-                return "morning"
-            if 12 <= hour < 18:
-                return "afternoon"
-            return parsed_datetime.strftime("%-I:%M %p")
-
-        if entry.get("is_daytime") is True:
-            return "daytime"
-        if entry.get("is_daytime") is False:
-            return "night"
-        return "forecast"
-
-    def _format_weather_forecast_entry(self, entry: Dict[str, Any]) -> str:
-        """Format a single weather forecast entry compactly."""
-        parts = []
-        condition = entry.get("condition")
-        if condition:
-            parts.append(str(condition).replace("_", " "))
-
-        high_temp = self._coerce_weather_temperature(entry.get("temperature"))
-        low_temp = self._coerce_weather_temperature(entry.get("templow"))
-        if high_temp is not None and low_temp is not None:
-            parts.append(
-                f"high {self._format_weather_temperature(high_temp)}, low {self._format_weather_temperature(low_temp)}"
-            )
-        elif high_temp is not None:
-            parts.append(f"around {self._format_weather_temperature(high_temp)}")
-
-        precipitation_probability = entry.get("precipitation_probability")
-        if precipitation_probability not in (None, ""):
-            parts.append(f"{precipitation_probability}% chance of rain")
-
-        wind_speed = entry.get("wind_speed")
-        if wind_speed not in (None, ""):
-            parts.append(f"wind {wind_speed}")
-
-        return ", ".join(parts) if parts else "forecast available"
-
-    @staticmethod
-    def _coerce_weather_temperature(value: Any) -> float | int | None:
-        """Coerce a weather temperature value safely."""
-        if value is None or value == "":
-            return None
-        if isinstance(value, (int, float)):
-            return value
-        try:
-            numeric = float(str(value).strip())
-        except (TypeError, ValueError):
-            return None
-        return int(numeric) if numeric.is_integer() else round(numeric, 1)
-
-    @staticmethod
-    def _format_weather_temperature(value: float | int) -> str:
-        """Format a compact temperature value."""
-        if isinstance(value, float) and value.is_integer():
-            value = int(value)
-        return str(value)
-
-    def _get_weather_target_forecast_types(
-        self, resolved_target: Dict[str, Any] | None
-    ) -> List[str]:
-        """Return forecast types supported by all targeted weather entities."""
-        entity_ids = self._normalize_target_values((resolved_target or {}).get("entity_id"))
-        if not entity_ids:
-            return []
-
-        common_types: set[str] | None = None
-        for entity_id in entity_ids:
-            state = self.hass.states.get(entity_id)
-            if state is None or state.domain != "weather":
-                continue
-
-            forecast_types = set(self.discovery._get_weather_forecast_types(state))
-            if not forecast_types:
-                continue
-
-            common_types = (
-                forecast_types
-                if common_types is None
-                else common_types.intersection(forecast_types)
-            )
-
-        if not common_types:
-            return []
-
-        return [
-            forecast_type
-            for forecast_type in ("daily", "twice_daily", "hourly")
-            if forecast_type in common_types
-        ]
-
-    def _select_preferred_weather_forecast_type(
-        self, forecast_types: List[str] | set[str] | tuple[str, ...]
-    ) -> str | None:
-        """Choose a stable forecast-type fallback for weather responses."""
-        forecast_type_set = {str(item) for item in forecast_types if item}
-        for candidate in ("daily", "twice_daily", "hourly"):
-            if candidate in forecast_type_set:
-                return candidate
-        return next(iter(sorted(forecast_type_set)), None)
-
-    async def _get_response_service_catalog(self) -> Dict[str, Dict[str, Dict[str, Any]]]:
-        """Build a catalog of HA services that support response data."""
-        descriptions = await service_helper.async_get_all_descriptions(self.hass)
-        catalog: Dict[str, Dict[str, Dict[str, Any]]] = {}
-
-        for domain, services in self.hass.services.async_services().items():
-            if self._get_domain_capability_error(domain):
-                continue
-
-            for service_name in services:
-                supports_response = self.hass.services.supports_response(
-                    domain, service_name
-                )
-                if supports_response == SupportsResponse.NONE:
-                    continue
-
-                description = dict(descriptions.get(domain, {}).get(service_name, {}))
-                description["supports_response"] = (
-                    "optional"
-                    if supports_response == SupportsResponse.OPTIONAL
-                    else "only"
-                )
-                catalog.setdefault(domain, {})[service_name] = description
-
-        return catalog
-
-    async def _get_response_service_info(
-        self, domain: str, service: str
-    ) -> tuple[Dict[str, Any], str | None]:
-        """Get dynamic metadata for a response-capable HA service."""
-        capability_error = self._get_domain_capability_error(domain)
-        if capability_error:
-            return {}, capability_error
-
-        if not self.hass.services.has_service(domain, service):
-            domain_info = get_domain_info(domain)
-            if domain_info is None:
-                return {}, (
-                    f"Domain '{domain}' is not registered in Home Assistant right now. "
-                    "Use list_response_services() to inspect available response-capable services."
-                )
-            return {}, (
-                f"Service '{domain}.{service}' is not registered in Home Assistant right now. "
-                "Use list_response_services() to inspect available response-capable services."
-            )
-
-        supports_response = self.hass.services.supports_response(domain, service)
-        if supports_response == SupportsResponse.NONE:
-            return {}, (
-                f"Service '{domain}.{service}' does not support native response data. "
-                "Use list_response_services() to find services that do."
-            )
-
-        catalog = await self._get_response_service_catalog()
-        description = dict(catalog.get(domain, {}).get(service, {}))
-        description["supports_response"] = (
-            "optional"
-            if supports_response == SupportsResponse.OPTIONAL
-            else "only"
-        )
-        return description, None
-
-    def _validate_response_service_parameters(
-        self, description: Dict[str, Any], provided_params: Dict[str, Any]
-    ) -> tuple[bool, str]:
-        """Validate required service parameters from HA's native service description."""
-        required_fields = self._get_required_service_fields(description)
-        missing = [field for field in required_fields if field not in provided_params]
-        if missing:
-            return (
-                False,
-                "Missing required parameters: " + ", ".join(missing),
-            )
-
-        return True, "Parameters valid"
-
-    def _get_required_service_fields(self, description: Dict[str, Any]) -> List[str]:
-        """Extract required field names from a service description."""
-        fields = description.get("fields", {})
-        if not isinstance(fields, dict):
-            return []
-
-        required_fields = []
-        for field_name, metadata in fields.items():
-            if isinstance(metadata, dict) and metadata.get("required") is True:
-                required_fields.append(str(field_name))
-
-        return sorted(required_fields)
-
-    def _get_service_field_names(self, description: Dict[str, Any]) -> List[str]:
-        """Extract service field names from a service description."""
-        fields = description.get("fields", {})
-        if not isinstance(fields, dict):
-            return []
-
-        return sorted(str(field_name) for field_name in fields.keys())
-
-    def _extract_service_target_domains(
-        self, description: Dict[str, Any]
-    ) -> List[str]:
-        """Extract allowed target entity domains from a service description."""
-        target = description.get("target")
-        if not isinstance(target, dict):
-            return []
-
-        entity_target = target.get("entity")
-        if not isinstance(entity_target, dict):
-            return []
-
-        domain_value = entity_target.get("domain")
-        if domain_value is None:
-            return []
-        if isinstance(domain_value, str):
-            return [domain_value]
-        if isinstance(domain_value, (list, tuple, set)):
-            return [str(item) for item in domain_value if item]
-        return [str(domain_value)]
-
-    def _restrict_resolved_target_for_service(
-        self,
-        resolved_target: Dict[str, Any],
-        *,
-        default_domain: str | None = None,
-        service_description: Dict[str, Any] | None = None,
-    ) -> Dict[str, Any]:
-        """Restrict resolved entity targets using known service target metadata."""
-        allowed_domains = []
-        if service_description:
-            allowed_domains = self._extract_service_target_domains(service_description)
-        if not allowed_domains and default_domain:
-            allowed_domains = [default_domain]
-
-        entity_ids = self._normalize_target_values(resolved_target.get("entity_id"))
-        if not allowed_domains:
-            resolved_domains = {entity_id.split(".", 1)[0] for entity_id in entity_ids}
-            if len(resolved_domains) > 1:
-                raise ValueError(
-                    "Resolved target spans multiple domains, and this service does not "
-                    "publish target-domain metadata. Use explicit entity_id values from discovery."
-                )
-            return {"entity_id": entity_ids}
-
-        filtered_entity_ids = [
-            entity_id
-            for entity_id in entity_ids
-            if entity_id.split(".", 1)[0] in allowed_domains
-        ]
-
-        if not filtered_entity_ids:
-            raise ValueError(
-                "Resolved target did not include any exposed entities accepted by this service."
-            )
-
-        return {"entity_id": filtered_entity_ids}
-
-    def _restrict_resolved_target_to_domain(
-        self, resolved_target: Dict[str, Any], domain: str
-    ) -> Dict[str, Any]:
-        """Restrict resolved entity targets to the requested domain."""
-        return self._restrict_resolved_target_for_service(
-            resolved_target, default_domain=domain
-        )
-
-    def _serialize_service_response_value(self, value: Any) -> Any:
-        """Serialize HA service response data to JSON-safe values."""
-        if value is None or isinstance(value, (str, int, float, bool)):
-            return value
-        if isinstance(value, (datetime, date, time)):
-            return value.isoformat()
-        if isinstance(value, dict):
-            return {
-                str(key): self._serialize_service_response_value(item)
-                for key, item in value.items()
-            }
-        if isinstance(value, (list, tuple, set)):
-            return [self._serialize_service_response_value(item) for item in value]
-        if isinstance(value, bytes):
-            try:
-                return value.decode("utf-8")
-            except UnicodeDecodeError:
-                return value.hex()
-        return str(value)
-
-    def _format_service_response_result(
-        self,
-        domain: str,
-        service: str,
-        resolved_target: Dict[str, Any],
-        response: Any,
-        *,
-        request_data: Dict[str, Any] | None = None,
-    ) -> str:
-        """Format a response-returning service call for the LLM."""
-        entity_ids = self._normalize_target_values(resolved_target.get("entity_id"))
-        if entity_ids:
-            target_count = len(entity_ids)
-            target_label = "entity" if target_count == 1 else "entities"
-            header = (
-                f"✅ Retrieved response from {domain}.{service} for "
-                f"{target_count} target {target_label}."
-            )
-        else:
-            header = f"✅ Retrieved response from {domain}.{service}."
-
-        text_parts = [header]
-
-        if domain == "weather" and service in {"get_forecast", "get_forecasts"}:
-            forecast_type = str((request_data or {}).get("type") or "").strip()
-            if forecast_type:
-                text_parts.append("")
-                text_parts.append(f"Forecast type used: {forecast_type}.")
-
-        summary_lines = self._build_service_response_summary(domain, service, response)
-        if summary_lines:
-            text_parts.append("")
-            text_parts.extend(summary_lines)
-
-        if response is None:
-            text_parts.append("")
-            text_parts.append("No response data was returned.")
-        else:
-            text_parts.append("")
-            text_parts.append("Response:")
-            text_parts.append(json.dumps(response, indent=2))
-
-        return "\n".join(text_parts)
-
-    def _build_service_response_summary(
-        self, domain: str, service: str, response: Any
-    ) -> List[str]:
-        """Build a concise summary for structured service responses."""
-        if domain == "music_assistant":
-            return summarize_music_assistant_response(service, response)
-        if domain == "weather" and service in {"get_forecast", "get_forecasts"}:
-            return self._summarize_weather_response(response)
-        if domain == "calendar" and service == "get_events":
-            return self._summarize_calendar_response(response)
-
-        if isinstance(response, dict):
-            return [f"Summary: {len(response)} top-level response entries."]
-        if isinstance(response, list):
-            return [f"Summary: {len(response)} response items."]
-        return []
-
-    def _summarize_weather_response(self, response: Any) -> List[str]:
-        """Summarize weather forecast response data."""
-        if not isinstance(response, dict):
-            return []
-
-        lines = ["Summary:"]
-        for entity_id, payload in response.items():
-            forecast = payload.get("forecast") if isinstance(payload, dict) else None
-            if not isinstance(forecast, list):
-                lines.append(f"- {entity_id}: no forecast entries returned")
-                continue
-
-            detail_parts = [f"{len(forecast)} forecast entries"]
-            if forecast:
-                first_forecast = forecast[0] if isinstance(forecast[0], dict) else {}
-                preview_parts = []
-                preview_time = self._format_service_response_datetime(
-                    first_forecast.get("datetime")
-                )
-                if preview_time:
-                    preview_parts.append(preview_time)
-                if first_forecast.get("condition"):
-                    preview_parts.append(str(first_forecast["condition"]))
-                high_temp = first_forecast.get("temperature")
-                low_temp = first_forecast.get("templow")
-                if high_temp is not None and low_temp is not None:
-                    preview_parts.append(f"{high_temp}/{low_temp}")
-                elif high_temp is not None:
-                    preview_parts.append(str(high_temp))
-                if preview_parts:
-                    detail_parts.append("first: " + ", ".join(preview_parts))
-
-            lines.append(f"- {entity_id}: {'; '.join(detail_parts)}")
-
-        return lines if len(lines) > 1 else []
-
-    def _summarize_calendar_response(self, response: Any) -> List[str]:
-        """Summarize calendar event response data."""
-        if not isinstance(response, dict):
-            return []
-
-        lines = ["Summary:"]
-        for entity_id, payload in response.items():
-            events = payload.get("events") if isinstance(payload, dict) else None
-            if not isinstance(events, list):
-                lines.append(f"- {entity_id}: no events returned")
-                continue
-
-            detail_parts = [f"{len(events)} events"]
-            if events:
-                first_event = events[0] if isinstance(events[0], dict) else {}
-                event_summary = first_event.get("summary") or first_event.get("title")
-                event_start = self._extract_calendar_event_start(first_event)
-                preview_parts = []
-                if event_summary:
-                    preview_parts.append(str(event_summary))
-                if event_start:
-                    preview_parts.append(event_start)
-                if preview_parts:
-                    detail_parts.append("next: " + " at ".join(preview_parts[:2]))
-
-            lines.append(f"- {entity_id}: {'; '.join(detail_parts)}")
-
-        return lines if len(lines) > 1 else []
-
-    def _extract_calendar_event_start(self, event: Dict[str, Any]) -> str | None:
-        """Extract and format the start time from a calendar event payload."""
-        if not isinstance(event, dict):
-            return None
-
-        start_value = event.get("start")
-        if isinstance(start_value, dict):
-            start_value = (
-                start_value.get("dateTime")
-                or start_value.get("datetime")
-                or start_value.get("date")
-            )
-
-        return self._format_service_response_datetime(start_value)
-
-    def _format_service_response_datetime(self, value: Any) -> str | None:
-        """Format a service-response date/time value in local time when possible."""
-        if value is None:
-            return None
-
-        if isinstance(value, datetime):
-            parsed = value
-        elif isinstance(value, date):
-            return value.isoformat()
-        else:
-            parsed = dt_util.parse_datetime(str(value))
-            if parsed is None:
-                return str(value)
-
-        if parsed.tzinfo is None:
-            parsed = parsed.replace(
-                tzinfo=getattr(dt_util, "DEFAULT_TIME_ZONE", dt_util.now().tzinfo)
-            )
-
-        return self._format_absolute_time(dt_util.as_utc(parsed))
 
     async def resolve_target(self, target: Dict[str, Any]) -> Dict[str, Any]:
         """Resolve target selectors to exposed entity IDs."""
