+++ custom_components/mcp_assist/server_tools/__init__.py
@@ -1,0 +1,1 @@
+"""Domain-specific MCP server tool mixins."""
